### 2023-10-30 #144 @NickleDave
Hi @vnmabus, just letting you know we have started the search for an editor for this review.

In the meantime, here are the initial editor checks. I am happy to report that rdata passes with flying colors.

Please see a couple of comments below. You are not required to address these for us to start the review, but I do think they would benefit your package.

---
---

Two suggestions:
1. include data in the package (more on this below)
2. it's great that you have examples, but add more concrete examples, e.g. using existing datasets on OSF, FigShare, or Zenodo. This will help your users understand how rdata fits into their workflow

More on adding data:
I tried to run the snippet in the README but got this error.

>>> parsed = rdata.parser.parse_file(rdata.TESTDATA_PATH / "test_vector.rda")
  File "/home/pimienta/Documents/repos/coding/opensci/pyos/test-rdata/.venv/lib/python3.10/site-packages/rdata/parser/_parser.py", line 1002, in parse_file
    data = path.read_bytes()
  File "/home/linuxbrew/.linuxbrew/opt/python@3.10/lib/python3.10/pathlib.py", line 1126, in read_bytes
    with self.open(mode='rb') as f:
  File "/home/linuxbrew/.linuxbrew/opt/python@3.10/lib/python3.10/pathlib.py", line 1119, in open
    return self._accessor.open(self, mode, buffering, encoding, errors,
FileNotFoundError: [Errno 2] No such file or directory: '/home/pimienta/Documents/repos/coding/opensci/pyos/test-rdata/.venv/lib/python3.10/site-packages/rdata/tests/data/test_vector.rda'

```
I think what's going on is that the [test data](https://github.com/vnmabus/rdata/tree/develop/rdata/tests/data) is not in the built package?

You may have intended this snippet to be run with just the development version, but I would strongly suggest you add small amounts of data to the package itself. It's also possible you meant to "include" the data in your built package, but the setuptools build is not configured correctly? I have found the way setuptools does this a bit confusing in the past, but I think it's now possible for you to [include all the files *without* using a MANIFEST.in file](https://setuptools.pypa.io/en/latest/userguide/datafiles.html) (which if I understand correctly is now a "legacy" way of including files with setuptools that should no longer be used). Note that other build backends have their own methods for this; for example [flit would include everything by default, although you can have more fine grained control](https://flit.pypa.io/en/stable/pyproject_toml.html?highlight=include#sdist-section).

You can then provide access to this data to users (and yourself, for tests) by using importlib-resources as discussed in [this talk](https://www.youtube.com/watch?v=ZsGFU2qh73E). I have a draft section of the guide on how to do this [here](https://github.com/pyOpenSci/python-package-guide/pull/110/commits/277c5769d215d407c56746d3856d1822a560d605)--it's very rough still but the core ideas are there and might help you. If you have any feedback on this section, we'd love to hear it.
