+++ rdata/__init__.py
@@ -1,12 +1,12 @@
 """rdata: Read R datasets from Python."""
-import os as _os
-import pathlib as _pathlib
+from importlib.resources import files
+from importlib.resources.abc import Traversable
 
 from . import conversion, parser
 
 
-def _get_test_data_path() -> _pathlib.Path:
-    return _pathlib.Path(_os.path.dirname(__file__)) / "tests" / "data"
+def _get_test_data_path() -> Traversable:
+    return files(__name__) / "tests" / "data"
 
 
 TESTDATA_PATH = _get_test_data_path()
