+++ rdata/conversion/_conversion.py
@@ -9,7 +9,7 @@ from typing import (
     Any,
     Callable,
     ChainMap,
-    List,
+    Final,
     Mapping,
     MutableMapping,
     NamedTuple,
@@ -33,14 +33,14 @@ StrMap = Mapping[Union[str, bytes], Any]
 class RLanguage(NamedTuple):
     """R language construct."""
 
-    elements: List[Any]
+    elements: list[Any]
     attributes: Mapping[str, Any]
 
 
 class RExpression(NamedTuple):
     """R expression."""
 
-    elements: List[RLanguage]
+    elements: list[RLanguage]
 
 
 @dataclass
@@ -96,7 +96,7 @@ class REnvironment(ChainMap[Union[str, bytes], Any]):
 def convert_list(
     r_list: parser.RObject,
     conversion_function: ConversionFunction,
-) -> Union[StrMap, List[Any]]:
+) -> StrMap | list[Any]:
     """
     Expand a tagged R pairlist to a Python dictionary.
 
@@ -206,8 +206,8 @@ def convert_attrs(
 def convert_vector(
     r_vec: parser.RObject,
     conversion_function: ConversionFunction,
-    attrs: Optional[StrMap] = None,
-) -> Union[List[Any], StrMap]:
+    attrs: StrMap | None = None,
+) -> list[Any] | StrMap:
     """
     Convert a R vector to a Python list or dictionary.
 
@@ -243,7 +243,7 @@ def convert_vector(
     }:
         raise TypeError("Must receive a VEC or EXPR object")
 
-    value: Union[List[Any], StrMap] = [
+    value: list[Any] | StrMap = [
         conversion_function(o) for o in r_vec.value
     ]
 
@@ -268,9 +268,9 @@ def safe_decode(byte_str: bytes, encoding: str) -> Union[str, bytes]:
 
 def convert_char(
     r_char: parser.RObject,
-    default_encoding: Optional[str] = None,
+    default_encoding: str | None = None,
     force_default_encoding: bool = False,
-) -> Union[str, bytes, None]:
+) -> str | bytes | None:
     """
     Decode a R character array to a Python string or bytes.
 
@@ -331,7 +331,7 @@ def convert_char(
 def convert_symbol(
     r_symbol: parser.RObject,
     conversion_function: ConversionFunction,
-) -> Union[str, bytes]:
+) -> str | bytes:
     """
     Decode a R symbol to a Python string or bytes.
 
@@ -364,8 +364,8 @@ def convert_symbol(
 def convert_array(
     r_array: RObject,
     conversion_function: ConversionFunction,
-    attrs: Optional[StrMap] = None,
-) -> Union[np.ndarray, xarray.DataArray]:
+    attrs: StrMap | None = None,
+) -> np.ndarray | xarray.DataArray:
     """
     Convert a R array to a Numpy ndarray or a Xarray DataArray.
 
@@ -573,7 +573,7 @@ ConstructorDict = Mapping[
     Constructor,
 ]
 
-default_class_map_dict: Mapping[Union[str, bytes], Constructor] = {
+default_class_map_dict: Final[Mapping[Union[str, bytes], Constructor]] = {
     "data.frame": dataframe_constructor,
     "factor": factor_constructor,
     "ordered": ordered_constructor,
@@ -583,7 +583,7 @@ default_class_map_dict: Mapping[Union[str, bytes], Constructor] = {
     "srcfilecopy": srcfilecopy_constructor,
 }
 
-DEFAULT_CLASS_MAP = MappingProxyType(default_class_map_dict)
+DEFAULT_CLASS_MAP: Final = MappingProxyType(default_class_map_dict)
 """
 Default mapping of constructor functions.
 
[…]
