+++ rdata/parser/_parser.py
@@ -9,19 +9,18 @@ import os
 import pathlib
 import warnings
 import xdrlib
+from collections.abc import Iterator
 from dataclasses import dataclass
 from types import MappingProxyType
 from typing import (
     Any,
     BinaryIO,
     Callable,
-    List,
+    Final,
     Mapping,
     Optional,
     Sequence,
-    Set,
     TextIO,
-    Tuple,
 )
 
 import numpy as np
@@ -31,7 +30,6 @@ try:
         Traversable as Traversable,
     )
 except ImportError:
-    from collections.abc import Iterator
     from typing import Protocol, runtime_checkable
 
     @runtime_checkable
@@ -96,7 +94,7 @@ magic_dict = {
 }
 
 
-def file_type(data: memoryview) -> Optional[FileTypes]:
+def file_type(data: memoryview) -> FileTypes | None:
     """Return the type of the file."""
     for filetype, magic in magic_dict.items():
         if data[:len(magic)] == magic:
@@ -112,14 +110,14 @@ class RdataFormats(enum.Enum):
     binary = "binary"
 
 
-format_dict = {
+format_dict: Final = MappingProxyType({
     RdataFormats.XDR: b"X\n",
     RdataFormats.ASCII: b"A\n",
     RdataFormats.binary: b"B\n",
-}
+})
 
 
-def rdata_format(data: memoryview) -> Optional[RdataFormats]:
+def rdata_format(data: memoryview) -> RdataFormats | None:
     """Return the format of the data."""
     for format_type, magic in format_dict.items():
         if data[:len(magic)] == magic:
@@ -166,7 +164,7 @@ class RObjectType(enum.Enum):
     REF = 255  # Reference
 
 
-BYTECODE_SPECIAL_SET = frozenset((
+BYTECODE_SPECIAL_SET: Final = frozenset((
     RObjectType.BCODE,
     RObjectType.BCREPREF,
     RObjectType.BCREPDEF,
@@ -224,7 +222,7 @@ class RObjectInfo():
 def _str_internal(
     obj: RObject | Sequence[RObject],
     indent: int = 0,
-    used_references: Optional[Set[int]] = None,
+    used_references: Optional[set[int]] = None,
 ) -> str:
 
     if used_references is None:
@@ -357,7 +355,7 @@ class EnvironmentValue():
 
 AltRepConstructor = Callable[
     [RObject],
-    Tuple[RObjectInfo, Any],
+    tuple[RObjectInfo, Any],
 ]
 AltRepConstructorMap = Mapping[bytes, AltRepConstructor]
 
@@ -378,7 +376,7 @@ def format_float_with_scipen(number: float, scipen: int) -> bytes:
 
 def deferred_string_constructor(
     state: RObject,
-) -> Tuple[RObjectInfo, Any]:
+) -> tuple[RObjectInfo, Any]:
     """Expand a deferred string ALTREP."""
     new_info = RObjectInfo(
         type=RObjectType.STR,
@@ -417,7 +415,7 @@ def compact_seq_constructor(
     state: RObject,
     *,
     is_int: bool = False,
-) -> Tuple[RObjectInfo, Any]:
+) -> tuple[RObjectInfo, Any]:
     """Expand a compact_seq ALTREP."""
     new_info = RObjectInfo(
         type=RObjectType.INT if is_int else RObjectType.REAL,
@@ -444,21 +442,21 @@ def compact_seq_constructor(
 
 def compact_intseq_constructor(
     state: RObject,
-) -> Tuple[RObjectInfo, Any]:
+) -> tuple[RObjectInfo, Any]:
     """Expand a compact_intseq ALTREP."""
     return compact_seq_constructor(state, is_int=True)
 
 
 def compact_realseq_constructor(
     state: RObject,
[…]
