### 2023-10-31 #144 @vnmabus
@NickleDave I have added the test data to the manifest and changed the package to use `importlib.resources` in the develop branch (but I have not yet released a version with the changes).

Some observations of the process:
- The `Traversable` protocol used by `importlib.resources` was not available in Python<3.11 so I had to redefine it in that case.
- I made `parse_file` able to load objects implementing the `Traversable` protocol. I think most Python developers are not aware of that interface, and thus most code is not capable to deal with general resources that are not in a filesystem. This probably should be discussed more widely in the Python community.
- I am not really sure that there is a more modern alternative to MANIFEST.in, so I keep using it. Thank you for noticing that the example data was missing!
