### 2023-11-01 2deebc9
Relax accepted types in parse_file.
