### 2023-11-02 b2e85b7
Improve NA handling for logical and integer types.
