### 2023-11-06 a253d99
Add example of loading from CRAN.
