+++ rdata/__init__.py
@@ -18,4 +18,4 @@ Path of the test data.
 
 """
 
-__version__ = "0.10.0"
+__version__ = "0.10.1.dev0"
