### 2023-11-18 #144 @isabelizimm
:wave: Hi @rich-iannone and @has2k1! Thank you SO MUCH for volunteering to review for pyOpenSci 🎉

@vnmabus these two reviewers are individuals who are deeply involved in both the R and Python data and packaging world. I am super excited to be part of and learn from them through this review process!

If anyone has any questions/comments/concerns, please don't hesitate to get in touch! Your review is due: Dec 8
