### 2023-11-20 147d8df
Fix bug in interpretation of compact_seq
