### 2023-11-21 26be6ef
Add faster reader for int arrays
