### 2023-11-21 285f707
Use file-like object to simplify reading
