### 2023-11-21 330a74b
Fix mypy and flake8 issues
