### 2023-11-22 9b4e578
Move xdr parser to a separate file
