+++ rdata/parser/_parser.py
@@ -14,7 +14,6 @@ from dataclasses import dataclass
 from types import MappingProxyType
 from typing import (
     Any,
-    BinaryIO,
     Callable,
     Final,
     Mapping,
@@ -997,79 +996,6 @@ class Parser(abc.ABC):
         return result
 
 
-class ParserXDR(Parser):
-    """Parser used when the integers and doubles are in XDR format."""
-
-    def __init__(
-        self,
-        file: BinaryIO,
-        *,
-        expand_altrep: bool = True,
-        altrep_constructor_dict: AltRepConstructorMap = DEFAULT_ALTREP_MAP,
-    ) -> None:
-        super().__init__(
-            expand_altrep=expand_altrep,
-            altrep_constructor_dict=altrep_constructor_dict,
-        )
-        self.file = file
-
-    def _parse_array(self,
-                     itemkind: str,
-                     itemsize: int,
-                     *,
-                     length: int | None = None,
-    ) -> npt.NDArray[Any]:  # noqa: D102
-        if length is None:
-            length = self.parse_int()
-
-        itemtype = f'{itemkind}{itemsize}'
-        buffer = self.file.read(length * itemsize)
-        # Read in big-endian order and convert to native byte order
-        return np.frombuffer(buffer, dtype=f'>{itemtype}').astype(f'={itemtype}')
-
-    def parse_int(self) -> int:  # noqa: D102
-        return int(self._parse_array('i', 4, length=1)[0])
-
-    def parse_double(self) -> float:  # noqa: D102
-        return float(self._parse_array('f', 8, length=1)[0])
-
-    def parse_complex(self) -> complex:  # noqa: D102
-        return complex(self._parse_array('c', 16, length=1)[0])
-
-    def parse_nullable_int_array(
-        self,
-        fill_value: int = R_INT_NA,
-    ) -> npt.NDArray[np.int32] | np.ma.MaskedArray[Any, Any]:  # noqa: D102
-
-        data = self._parse_array('i', 4)
-        mask = data == R_INT_NA
-        data[mask] = fill_value
-
-        if np.any(mask):
-            return np.ma.array(  # type: ignore
-                data=data,
-                mask=mask,
-                fill_value=fill_value,
-            )
-
-        return data
-
-    def parse_double_array(self) -> npt.NDArray[np.float64]:  # noqa: D102
-        return self._parse_array('f', 8)
-
-    def parse_complex_array(self) -> npt.NDArray[np.complex128]:  # noqa: D102
-        return self._parse_array('c', 16)
-
-    def parse_string(self, length: int) -> bytes:  # noqa: D102
-        return self.file.read(length)
-
-    def parse_all(self) -> RData:
-        rdata = super().parse_all()
-        # Check that there is no more data in the file
-        assert self.file.read(1) == b''
-        return rdata
-
-
 def parse_file(
     file_or_path: AcceptableFile | os.PathLike[Any] | Traversable | str,
     *,
@@ -1322,6 +1248,8 @@ def parse_rdata_binary(
         data = data[len(format_dict[format_type]):]
 
     if format_type is RdataFormats.XDR:
+        from rdata.io.xdr import ParserXDR
+
         parser = ParserXDR(
             io.BytesIO(data),
             expand_altrep=expand_altrep,
