### 2023-11-22 b44b7d5
Avoid unnecessary copy
