### 2023-11-27 846817e
Assert that there is no unexpected value
