### 2023-11-27 ee7e635
Simplify string ndarray handling
