### 2023-11-27 a17e1ad
Add converter for dict
