### 2023-11-27 816a2a4
Add function for building r lists
