### 2023-11-27 7a87f0f
Convert arrays of any dimensionality
