### 2023-11-28 a34f71c
Use functions instead of a class
