### 2023-11-28 bbdb899
Add docstrings
