+++ rdata/conversion/__init__.py
@@ -24,3 +24,8 @@ from ._conversion import (
     factor_constructor as factor_constructor,
     ts_constructor as ts_constructor,
 )
+
+from .to_r import (
+    convert_to_r_data,
+    convert_to_r_object,
+)
