+++ rdata/io/__init__.py
@@ -1,4 +1,30 @@
-def write(fpath, data, *, format, compression='gzip'):
+from __future__ import annotations
+
+import os
+from typing import Any
+
+
+def write(
+        path: os.PathLike[Any] | str,
+        r_data: RData,
+        *,
+        format: str = 'xdr',
+        compression: str = 'gzip',
+) -> None:
+    """
+    Write RData object to a file.
+
+    Parameters
+    ----------
+    path:
+        File path to be written
+    r_data:
+        RData object
+    format:
+        File format (ascii or xdr)
+    compression:
+        Compression (gzip, bzip2, xz, or none)
+    """
     if format == 'ascii':
         from .ascii import WriterASCII as Writer
         mode = 'w'
@@ -14,12 +40,15 @@ def write(fpath, data, *, format, compression='gzip'):
         from bz2 import open
     elif compression == 'xz':
         from lzma import open
-    else:
-        assert compression is None
+    elif compression == 'none':
         import builtins
         open = builtins.open
+    else:
+        msg = f'Unknown compression: {compression}'
+        if compression is None:
+            msg += ". Use 'none' for no compression."
+        raise ValueError(msg)
 
-    with open(fpath, mode) as f:
+    with open(path, mode) as f:
         w = Writer(f)
-        w.write_all(data)
-
+        w.write_all(r_data)
