### 2023-11-28 ac3fde5
Rename write_all() to write_r_data()
