### 2023-12-04 074a825
Use dtype object instead of itemkind and itemsize
