### 2023-12-04 afb0d0e
Clarify array value reading
