### 2023-12-08 #144 @has2k1
> That said, it would be better if you include in your review also the recent changes in develop

That is okay with me.

On `xdrlib`. I had seen that it is deprecated and getting removed and then thankfully a PR showed up. I think, eventually the best place for the parser is as a separate package.
