### 2023-12-11 #144 @has2k1
---

##### Notes & Suggestions

1. For the code in the README (and since as it is not a doctest), it would be convenient to be able to copy & paste it and run without any editing.

    - The default python kernel does ignore `>>>` & `...`, though jupyter kernel does.
    - Having the output in the same code block breaks the copy-paste-run for all kernels.

2. I feel like there is are missing `read` functions that wraps `rdata.parser.parse_file` and
`rdata.parser.convert` and works with just a filepath. e.g.

  ```py
  parsed = rdata.parser.parse_file("dataframe.rds")
  data = rdata.conversion.convert(parsed)
  ```

  would be

  ```py
  data = rdata.read_rds("dataframe.rds")
  ```

  And maybe also a `read_rda`. A single `read_rdata` can do for both cases but that does not communicate the difference between reading an object vs an environment with objects.

  These functions could include optional arguments to pass on to the _parser_ and/or _converter_.

  This would simplify the user experience as the `convert(parse_file(...))` usage seems to cover majority of the use cases. e.g. none of the current test cases deviate from that pattern.

3. It is not documented anywhere what base R objects are supported and what they are translate to as python object. I think a table would be helpful.

4. On packaging configuration, there is a `setup.cfg` and `pyproject.toml`. As even the configuration for `setuptools` is in `pyproject.toml`, it makes sense to migrate the configs for `isort`, `mypy` and `pytest`. That would leave only the `flake8` configuration in `setup.cfg` as `flake8` [does not yet support](https://github.com/PyCQA/flake8/issues/234) `pyproject.toml`.

5. `xdrlib` has been deprecated in python 3.12 and it is good there is a plan (and some work) to deal with it.

##### Final Comment

While I have noted some suggestions, the project is otherwise well motivated, implemented and tested and I do not consider the above suggestions to be "blockers".
