### 2023-12-12 b37aa9f
Add missing badges.
