### 2023-12-12 c73eb3c
Try fix tests in 3.12.
