### 2023-12-12 60aa43d
Add issue template for bug.
