+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,23 +1,0 @@
----
-name: Bug report
-about: Create a report to help us improve
-title: ''
-labels: bug
-assignees: ''
-
----
-
-**Describe the bug**
-A clear and concise description of what the bug is.
-
-**To Reproduce**
-Indicate the steps to reproduce the behavior.
-
-**Expected behavior**
-A clear and concise description of what you expected to happen.
-
-**Technical information**
- - OS: [e.g. Windows 10]
-
-**Additional context**
-Add any other context about the problem here.
