+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,8 +1,6 @@
 name: Bug report
 about: Create a report to help us reproduce and fix a bug
-title: ''
 labels: bug
-assignees: ''
 
 body:
 - type: markdown
