+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,5 +1,5 @@
 name: Bug report
-about: Create a report to help us reproduce and fix a bug
+description: Create a report to help us reproduce and fix a bug
 labels: bug
 
 body:
