### 2023-12-18 ab24a7e
Move xdr parser to parser directory
