### 2023-12-18 5104fa4
Remove duplicate definition of R_INT_NA
