### 2023-12-18 d2d318c
Pass memoryview to xdr parser
