### 2023-12-18 10ba6de
Declare _parse_array_values() as abstract
