+++ CONTRIBUTORS.md
@@ -1,0 +1,42 @@
+
+## Contributors ✨
+
+Thanks goes to these wonderful people ([emoji key](https://allcontributors.org/docs/en/emoji-key)):
+
+<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
+<!-- prettier-ignore-start -->
+<!-- markdownlint-disable -->
+<table>
+  <tbody>
+    <tr>
+      <td align="center" valign="top" width="14.28%"><a href="https://github.com/vnmabus"><img src="https://avatars.githubusercontent.com/u/2364173?v=4?s=100" width="100px;" alt="Carlos Ramos Carreño"/><br /><sub><b>Carlos Ramos Carreño</b></sub></a><br /><a href="https://github.com/VNMabus/rdata/commits?author=vnmabus" title="Code">💻</a> <a href="#data-vnmabus" title="Data">🔣</a> <a href="https://github.com/VNMabus/rdata/commits?author=vnmabus" title="Documentation">📖</a> <a href="#example-vnmabus" title="Examples">💡</a> <a href="#ideas-vnmabus" title="Ideas, Planning, & Feedback">🤔</a> <a href="#infra-vnmabus" title="Infrastructure (Hosting, Build-Tools, etc)">🚇</a> <a href="#maintenance-vnmabus" title="Maintenance">🚧</a> <a href="#projectManagement-vnmabus" title="Project Management">📆</a> <a href="#question-vnmabus" title="Answering Questions">💬</a> <a href="https://github.com/VNMabus/rdata/pulls?q=is%3Apr+reviewed-by%3Avnmabus" title="Reviewed Pull Requests">👀</a> <a href="https://github.com/VNMabus/rdata/commits?author=vnmabus" title="Tests">⚠️</a> <a href="#tutorial-vnmabus" title="Tutorials">✅</a></td>
+      <td align="center" valign="top" width="14.28%"><a href="https://www.csc.fi"><img src="https://upload.wikimedia.org/wikipedia/en/6/6e/CSC_%E2%80%93_IT_Center_for_Science_%28logo%29.jpg?s=100" width="100px;" alt="CSC - IT Center for Science Ltd"/><br /><sub><b>CSC - IT Center for Science Ltd</b></sub></a><br /><a href="https://github.com/vnmabus/rdata/commits?author=trossi" title="Code">💻</a></td>
+      <td align="center" valign="top" width="14.28%"><a href="https://github.com/trossi"><img src="https://avatars.githubusercontent.com/u/34502776?v=4?s=100" width="100px;" alt="Tuomas Rossi"/><br /><sub><b>Tuomas Rossi</b></sub></a><br /><a href="https://github.com/VNMabus/rdata/commits?author=trossi" title="Code">💻</a> <a href="#ideas-trossi" title="Ideas, Planning, & Feedback">🤔</a> <a href="https://github.com/VNMabus/rdata/issues?q=author%3Atrossi" title="Bug reports">🐛</a></td>
+      <td align="center" valign="top" width="14.28%"><a href="https://www.researchgate.net/profile/Vladimir_Vargas-Calderon"><img src="https://avatars.githubusercontent.com/u/31494271?v=4?s=100" width="100px;" alt="Vladimir Vargas-Calderón"/><br /><sub><b>Vladimir Vargas-Calderón</b></sub></a><br /><a href="https://github.com/VNMabus/rdata/issues?q=author%3AVolodyaCO" title="Bug reports">🐛</a></td>
[…]
