+++ pyproject.toml
@@ -11,7 +11,6 @@ keywords = [
 ]
 authors = [
   {name = "Carlos Ramos Carreño", email = "vnmabus@gmail.com"},
-  {name = "Tuomas Rossi", email = "tuomas.rossi@csc.fi"},
 ]
 maintainers = [
   {name = "Carlos Ramos Carreño", email = "vnmabus@gmail.com"},
