### 2024-01-17 cd0e4a8
Add issue template for features.
