+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,6 +1,6 @@
 name: Bug report
 description: Create a report to help us reproduce and fix a bug
-labels: bug
+labels: [bug]
 
 body:
 - type: markdown
