+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,31 @@
+<!--
+Thanks for your contribution! Please ensure you have taken a look at
+the contribution guidelines:
+https://github.com/VNMabus/rdata/blob/develop/CONTRIBUTING.md
+-->
+
+## References to issues or other PRs
+<!--
+Include links to the relevant issues and PRs, using the relevant Github
+keywords (e.g., Fixes) for closing automatically the issues resolved
+on merge (see https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue).
+If there is no current issue discussing the addition of this functionality, it
+is recommended to create one and discuss the feature there.
+Otherwise, it is possible for this functionality to be rejected, or to require
+considerable changes.
+Example: Fixes #42. See also #123.
+-->
+
+
+## Describe the proposed changes
+
+
+## Additional information
+
+
+## Checklist before requesting a review
+
+- [ ] I have performed a self-review of my code
+- [ ] The code conforms to the style used in this package
+- [ ] The code is fully documented and typed (type-checked with Mypy)
+- [ ] I have added thorough tests for the new/changed functionality
