### 2024-01-17 8fff3a9
Add Mypy workflow.
