+++ .github/workflows/mypy.yml
@@ -1,0 +1,19 @@
+name: Mypy
+
+on:
+  pull_request:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    name: Mypy
+    steps:
+    - uses: actions/checkout@v2
+    - uses: tsuyoshicho/action-mypy@v3
+      with:
+        github_token: ${{ secrets.github_token }}
+        reporter: github-pr-review
+        # The action will output fail if there are mypy errors
+        level: error
+        setup_command: pip install -r requirements.txt pytest mypy
+        mypy_flags: ''
