### 2024-01-17 a7fbcfa
Add test for ascii files
