+++ .github/workflows/mypy.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     name: Mypy
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v4
     - uses: tsuyoshicho/action-mypy@v3
       with:
         github_token: ${{ secrets.github_token }}
