### 2024-01-17 09a1e76
Some typing fixes.
