### 2024-01-17 dc306f5
Try to fix MyPy workflow.
