+++ rdata/parser/_parser.py
@@ -8,26 +8,16 @@ import lzma
 import os
 import pathlib
 import warnings
-from collections.abc import Iterator
+from collections.abc import Callable, Iterator, Mapping, Sequence
 from dataclasses import dataclass
 from types import MappingProxyType
-from typing import (
-    Any,
-    Callable,
-    Final,
-    Mapping,
-    Optional,
-    Protocol,
-    Sequence,
-    Union,
-    runtime_checkable,
-)
+from typing import Any, Final, Protocol, Union, runtime_checkable
 
 import numpy as np
 import numpy.typing as npt
 
 #: Value used to represent a missing integer in R.
-R_INT_NA: Final = -2**31  # noqa: WPS432
+R_INT_NA: Final = -2**31
 
 
 @runtime_checkable
@@ -50,16 +40,14 @@ class BinaryBufferFileLike(Protocol):
 AcceptableFile = Union[BinaryFileLike, BinaryBufferFileLike]
 
 try:
-    from importlib.resources.abc import (  # noqa:WPS113
-        Traversable as Traversable,
-    )
+    from importlib.resources.abc import Traversable as Traversable
 except ImportError:
 
     @runtime_checkable
     class Traversable(Protocol):  # type: ignore [no-redef]
         """Definition of Traversable protocol for Python < 3.11."""
 
-        def iterdir(self) -> Iterator["Traversable"]:
+        def iterdir(self) -> Iterator[Traversable]:
             pass
 
         def read_bytes(self) -> bytes:
@@ -77,20 +65,18 @@ except ImportError:
         def joinpath(
             self,
             *descendants: str | os.PathLike[str],
-        ) -> "Traversable":
+        ) -> Traversable:
             pass
 
         def __truediv__(
             self,
             child: str | os.PathLike[str],
-        ) -> "Traversable":
+        ) -> Traversable:
             pass
 
-        def open(
+        def open(  # noqa: A003
             self,
-            mode: str = 'r',
-            *args: Any,
-            **kwargs: Any,
+            mode: str = "r",
         ) -> AcceptableFile:
             pass
 
@@ -211,16 +197,16 @@ class CharFlags(enum.IntFlag):
 
 
 @dataclass
-class RVersions():
+class RVersions:
     """R versions."""
 
-    format: int  # noqa: E701
+    format: int  # noqa: A003
     serialized: int
     minimum: int
 
 
 @dataclass
-class RExtraInfo():
+class RExtraInfo:
     """
     Extra information.
 
@@ -228,25 +214,25 @@ class RExtraInfo():
 
     """
 
-    encoding: Optional[str] = None
+    encoding: str | None = None
 
 
 @dataclass
-class RObjectInfo():
+class RObjectInfo:
     """Internal attributes of a R object."""
 
-    type: RObjectType
-    object: bool
+    type: RObjectType  # noqa: A003
+    object: bool  # noqa: A003
     attributes: bool
     tag: bool
     gp: int
     reference: int
 
 
-def _str_internal(
+def _str_internal(  # noqa: PLR0912, C901
     obj: RObject | Sequence[RObject],
     indent: int = 0,
-    used_references: Optional[set[int]] = None,
+    used_references: set[int] | None = None,
 ) -> str:
 
     if used_references is None:
@@ -255,9 +241,9 @@ def _str_internal(
     small_indent = indent + 2
     big_indent = indent + 4
 
-    indent_spaces = ' ' * indent
-    small_indent_spaces = ' ' * small_indent
-    big_indent_spaces = ' ' * big_indent
+    indent_spaces = " " * indent
+    small_indent_spaces = " " * small_indent
+    big_indent_spaces = " " * big_indent
 
     string = ""
 
@@ -313,8 +299,10 @@ def _str_internal(
                 used_references.copy(),
[…]
