### 2024-01-18 af6e04b
Fix ruff problems in conversion module.
