+++ rdata/conversion/_conversion.py
@@ -2,32 +2,21 @@ from __future__ import annotations
 
 import abc
 import warnings
+from collections import ChainMap
+from collections.abc import Callable, Mapping, MutableMapping, Sequence
 from dataclasses import dataclass
 from fractions import Fraction
 from types import MappingProxyType, SimpleNamespace
-from typing import (
-    Any,
-    Callable,
-    ChainMap,
-    Final,
-    Mapping,
-    MutableMapping,
-    NamedTuple,
-    Optional,
-    Sequence,
-    Union,
-    cast,
-)
+from typing import Any, Final, NamedTuple, Union, cast
 
 import numpy as np
-import pandas
+import pandas as pd
 import xarray
+from typing_extensions import override
 
 from .. import parser
-from ..parser import RObject
 
 ConversionFunction = Callable[[Union[parser.RData, parser.RObject]], Any]
-StrMap = Mapping[Union[str, bytes], Any]
 
 
 class RLanguage(NamedTuple):
@@ -55,9 +44,9 @@ class RFunction:
     """R function."""
 
     environment: Mapping[str, Any]
-    formals: Optional[Mapping[str, Any]]
+    formals: Mapping[str, Any] | None
     body: RLanguage
-    attributes: StrMap
+    attributes: Mapping[str, Any]
 
     @property
     def source(self) -> str:
@@ -78,16 +67,16 @@ class RBytecode:
 
     code: xarray.DataArray
     constants: Sequence[Any]
-    attributes: StrMap
+    attributes: Mapping[str, Any]
 
 
-class REnvironment(ChainMap[Union[str, bytes], Any]):
+class REnvironment(ChainMap[str, Any]):
     """R environment."""
 
     def __init__(
         self,
-        *maps: MutableMapping[str | bytes, Any],
-        frame: StrMap | None = None,
+        *maps: MutableMapping[str, Any],
+        frame: Mapping[str, Any] | None = None,
     ) -> None:
         super().__init__(*maps)
         self.frame = frame
@@ -96,41 +85,37 @@ class REnvironment(ChainMap[Union[str, bytes], Any]):
 def convert_list(
     r_list: parser.RObject,
     conversion_function: ConversionFunction,
-) -> StrMap | list[Any]:
+) -> Mapping[str, Any] | list[Any]:
     """
     Expand a tagged R pairlist to a Python dictionary.
 
-    Parameters
-    ----------
-    r_list: RObject
-        Pairlist R object, with tags.
-    conversion_function: Callable
-        Conversion function to apply to the elements of the list. By default
-        is the identity function.
+    Parameters:
+        r_list: RObject
+            Pairlist R object, with tags.
+        conversion_function: Callable
+            Conversion function to apply to the elements of the list. By default
+            is the identity function.
 
-    Returns
-    -------
-    dictionary: dict
-        A dictionary with the tags of the pairwise list as keys and their
-        corresponding values as values.
+    Returns:
+        dictionary: dict
+            A dictionary with the tags of the pairwise list as keys and their
+            corresponding values as values.
 
-    See Also
-    --------
-    convert_vector
+    See Also:
+        convert_vector
 
     """
     if r_list.info.type is parser.RObjectType.NILVALUE:
         return {}
-    elif r_list.info.type not in {
+
+    if r_list.info.type not in {
         parser.RObjectType.LIST,
         parser.RObjectType.LANG,
     }:
-        raise TypeError("Must receive a LIST, LANG or NILVALUE object")
+        msg = "Must receive a LIST, LANG or NILVALUE object"
+        raise TypeError(msg)
 
-    if r_list.tag is None:
-        tag = None
-    else:
[…]
