+++ pyproject.toml
@@ -81,9 +81,14 @@ select = [
 ignore = [
 	"ANN101", # self does not need to be typed
 	"D212", # incompatible with D213, which is our preferred style for multiline docstrings
+	"Q003", # do not change quotation marks to avoid escaping
 	"PLC0414", # allow explicit re-exports
 	"S101", # assert is allowed
+	"TID252", # relative imports allowed
 ]
 
 [tool.ruff.pydocstyle]
 convention = "google"
+
+[tool.ruff.pylint]
+max-args = 7
