+++ rdata/conversion/__init__.py
@@ -1,3 +1,4 @@
+"""Utilities for converting R objects to Python ones."""
 from ._conversion import (
     DEFAULT_CLASS_MAP as DEFAULT_CLASS_MAP,
     Converter as Converter,
