+++ rdata/tests/test_rdata.py
@@ -4,7 +4,7 @@ import unittest
 from collections import ChainMap
 from fractions import Fraction
 from types import SimpleNamespace
-from typing import Any, Dict
+from typing import Any
 
 import numpy as np
 import pandas as pd
@@ -15,16 +15,16 @@ import rdata
 TESTDATA_PATH = rdata.TESTDATA_PATH
 
 
-class SimpleTests(unittest.TestCase):  # noqa:WPS214
+class SimpleTests(unittest.TestCase):
     """Collection of simple test cases."""
 
     def test_opened_file(self) -> None:
         """Test that an opened file can be passed to parse_file."""
-        with open(TESTDATA_PATH / "test_vector.rda") as f:
+        with (TESTDATA_PATH / "test_vector.rda").open() as f:
             parsed = rdata.parser.parse_file(f)
             converted = rdata.conversion.convert(parsed)
 
-            self.assertIsInstance(converted, dict)
+            assert isinstance(converted, dict)
 
     def test_opened_string(self) -> None:
         """Test that a string can be passed to parse_file."""
@@ -33,7 +33,7 @@ class SimpleTests(unittest.TestCase):  # noqa:WPS214
         )
         converted = rdata.conversion.convert(parsed)
 
-        self.assertIsInstance(converted, dict)
+        assert isinstance(converted, dict)
 
     def test_logical(self) -> None:
         """Test parsing of logical vectors."""
@@ -228,9 +228,9 @@ class SimpleTests(unittest.TestCase):  # noqa:WPS214
             "test_list":
                 [
                     np.array([1.0]),
-                    ['a', 'b', 'c'],
+                    ["a", "b", "c"],
                     np.array([2.0, 3.0]),
-                    ['hi'],
+                    ["hi"],
                 ],
         })
 
@@ -251,7 +251,7 @@ class SimpleTests(unittest.TestCase):  # noqa:WPS214
         np.testing.assert_equal(converted, {
             "test_expression": rdata.conversion.RExpression([
                 rdata.conversion.RLanguage(
-                    ['^', 'base', 'exponent'],
+                    ["^", "base", "exponent"],
                     attributes={},
                 ),
             ]),
@@ -275,7 +275,7 @@ class SimpleTests(unittest.TestCase):  # noqa:WPS214
 
         converted_fun = converted["test_minimal_function_uncompiled"]
 
-        self.assertIsInstance(
+        assert isinstance(
             converted_fun,
             rdata.conversion.RFunction,
         )
@@ -297,7 +297,7 @@ class SimpleTests(unittest.TestCase):  # noqa:WPS214
 
         converted_fun = converted["test_minimal_function"]
 
-        self.assertIsInstance(
+        assert isinstance(
             converted_fun,
             rdata.conversion.RFunction,
         )
@@ -307,7 +307,7 @@ class SimpleTests(unittest.TestCase):  # noqa:WPS214
 
         converted_body = converted_fun.body
 
-        self.assertIsInstance(
+        assert isinstance(
             converted_body,
             rdata.conversion.RBytecode,
         )
@@ -329,17 +329,17 @@ class SimpleTests(unittest.TestCase):  # noqa:WPS214
 
         converted_fun = converted["test_empty_function_uncompiled"]
 
-        self.assertIsInstance(
+        assert isinstance(
             converted_fun,
             rdata.conversion.RFunction,
         )
 
         np.testing.assert_equal(converted_fun.environment, ChainMap({}))
         np.testing.assert_equal(converted_fun.formals, None)
-        self.assertIsInstance(converted_fun.body, rdata.conversion.RLanguage)
+        assert isinstance(converted_fun.body, rdata.conversion.RLanguage)
         np.testing.assert_equal(
[…]
