+++ rdata/__init__.py
@@ -2,10 +2,12 @@
 from __future__ import annotations
 
 from importlib.resources import files
-from typing import Final
+from typing import TYPE_CHECKING, Final
 
-from . import conversion, parser
-from .parser._parser import Traversable
+from . import conversion as conversion, parser as parser
+
+if TYPE_CHECKING:
+    from .parser._parser import Traversable
 
 
 def _get_test_data_path() -> Traversable:
