+++ setup.cfg
@@ -139,6 +139,12 @@ strict = True
 strict_equality = True
 implicit_reexport = True
 
+[mypy-igraph.*]
+ignore_missing_imports = True
+
+[mypy-ipywidgets.*]
+ignore_missing_imports = True
+
 [mypy-numpy.*]
 ignore_missing_imports = True
 
@@ -147,3 +153,6 @@ ignore_missing_imports = True
 
 [mypy-setuptools.*]
 ignore_missing_imports = True
+
+[mypy-examples.*]
+disallow_untyped_defs = False
