### 2024-01-22 1bcbd1b
Rename Ruff sections and remove flake8.
