+++ pyproject.toml
@@ -75,7 +75,7 @@ include = ["rdata*"]
 [tool.setuptools.dynamic]
 version = {attr = "rdata.__version__"}
 
-[tool.ruff]
+[tool.ruff.lint]
 select = [
     "ALL",
 ]
@@ -88,7 +88,7 @@ ignore = [
 	"TID252", # relative imports allowed
 ]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "plot_*.py" = [
 	"ANN", # no type hints in examples
 	"ARG001", # Some unused args are needed
@@ -103,11 +103,11 @@ ignore = [
 	"SIM117", # multiple with necessary for now
 ]
 
-[tool.ruff.isort]
+[tool.ruff.lint.isort]
 combine-as-imports = true
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "google"
 
-[tool.ruff.pylint]
+[tool.ruff.lint.pylint]
 max-args = 7
