### 2024-01-22 e9b3d50
Fix Mypy errors and move Mypy config to pyproject.toml
