+++ setup.cfg
@@ -12,26 +12,3 @@ include_trailing_comma = true
 use_parentheses = true
 combine_as_imports = 1
 skip_glob = **/plot_*.py plot_*.py
-
-[mypy]
-strict = True
-strict_equality = True
-implicit_reexport = True
-
-[mypy-igraph.*]
-ignore_missing_imports = True
-
-[mypy-ipywidgets.*]
-ignore_missing_imports = True
-
-[mypy-numpy.*]
-ignore_missing_imports = True
-
-[mypy-pandas.*]
-ignore_missing_imports = True
-
-[mypy-setuptools.*]
-ignore_missing_imports = True
-
-[mypy-examples.*]
-disallow_untyped_defs = False
