### 2024-01-22 f23570b
Fix Mypy config.
