### 2024-01-22 e99dd49
Move Pytest settings to pyproject.toml
