+++ setup.cfg
@@ -1,11 +1,3 @@
-[aliases]
-test=pytest
-
-[tool:pytest]
-addopts = --doctest-modules --doctest-glob="*.rst"
-doctest_optionflags = NORMALIZE_WHITESPACE ELLIPSIS
-norecursedirs = .* build dist *.egg venv .svn _build docs/auto_examples examples
-
 [isort]
 multi_line_output = 3
 include_trailing_comma = true
