### 2024-01-22 2b927cb
Move isort config to pyproject.toml
