+++ setup.cfg
@@ -1,6 +1,0 @@
-[isort]
-multi_line_output = 3
-include_trailing_comma = true
-use_parentheses = true
-combine_as_imports = 1
-skip_glob = **/plot_*.py plot_*.py
