### 2024-01-22 696d95b
Declare mypy error code
