### 2024-01-22 af24a29
Clarify string encoding
