+++ rdata/parser/_ascii.py
@@ -62,9 +62,39 @@ class ParserASCII(Parser):
         return array
 
     def parse_string(self, length: int) -> bytes:
-        s = self._readline().encode('ascii').decode('unicode_escape').encode('latin1')
-        assert len(s) == length
-        return s
+        # Non-ascii characters in strings are written using octal byte codes,
+        # for example, a string 'aä' (2 chars) in UTF-8 is written as an ascii
+        # string r'a\303\244' (9 chars). We want to transform this to a byte
+        # string b'a\303\244' (3 bytes) corresponding to the byte
+        # representation of the original UTF-8 string.
+        # Let's use this string as an example to go through the code below
+
+        # Read the ascii string
+        s = self._readline()
+        # Now s = r'a\303\244' (9 chars)
+
+        # Convert characters to bytes (all characters are ascii)
+        b = s.encode('ascii')
+        # Now b = br'a\303\244' (9 bytes)
+
+        # There is a special 'unicode_escape' encoding that does
+        # basically two things here:
+        # 1) interpret e.g. br'\303' (4 bytes) as a single byte b'\303'
+        # 2) decode so-transformed byte string to a string with latin1 encoding
+        s = b.decode('unicode_escape')
+        # Now s = 'aÃ¤' (3 chars)
+
+        # We don't really want the latter latin1 decoding step done by
+        # the previous line of code, so we undo it by encoding in latin1
+        # back to bytes
+        b = s.encode('latin1')
+        # Now b = b'a\303\244' (3 bytes)
+
+        # We return this byte representation here. Later in the code there
+        # will be the decoding step from b'a\303\244' to 'aä', that is,
+        # s = b.decode('utf8')
+        assert len(b) == length
+        return b
 
     def check_complete(self) -> None:
         assert self.file.read(1) == ''
