### 2024-01-22 582f8eb
Add parser for ascii files
