### 2024-01-22 66b8444
Add support for ascii files generated on Windows
