### 2024-01-22 72a1ec4
Remove duplicated code across parsers
