### 2024-01-22 16125d9
Fix mypy errors
