### 2024-01-22 0f7d4c1
Add type hint for parser class
