### 2024-01-23 vnmabus/rdata#35 @vnmabus
Improve tests
There are a couple of ideas for improving the tests in this package.

For the tests, I think that the ideal thing would be to have each dataset as both binary and ASCII. Then it should be possible, maybe using a Pytest fixture, to pass the same tests for both versions, ensuring that all functionality work for both implementations.

_Originally posted by @vnmabus in https://github.com/vnmabus/rdata/pull/34#discussion_r1455873764_

I think it could also be useful to include in the repo an R script that generates all those test data files.

_Originally posted by @trossi in https://github.com/vnmabus/rdata/pull/34#discussion_r1462228483_
