### 2024-01-23 1500547
Update MyPy action version.
