+++ .github/workflows/mypy.yml
@@ -9,7 +9,7 @@ jobs:
     name: Mypy
     steps:
     - uses: actions/checkout@v4
-    - uses: tsuyoshicho/action-mypy@v3
+    - uses: tsuyoshicho/action-mypy@v4
       with:
         github_token: ${{ secrets.github_token }}
         reporter: github-pr-review
