+++ .github/workflows/ruff.yml
@@ -6,3 +6,5 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - uses: chartboost/ruff-action@v1
+        with:
+          args: check --output-format github
