### 2024-01-23 ef10690
Do not execute Rust twice.
