### 2024-01-23 825bedb
Fix Ruff error.
