### 2024-01-23 28cbdc7
Fix Mypy action.
