+++ .github/workflows/mypy.yml
@@ -9,11 +9,12 @@ jobs:
     name: Mypy
     steps:
     - uses: actions/checkout@v4
+    - name: Install dependencies
+      run: |
+        pip install ".[test,typing]" mypy;
     - uses: tsuyoshicho/action-mypy@v4
       with:
         github_token: ${{ secrets.github_token }}
         reporter: github-pr-review
         # The action will output fail if there are mypy errors
         level: error
-        setup_command: pip install ".[test,typing]" mypy
-        mypy_flags: ''
