### 2024-01-23 4972eda
Check Mypy reporting.
