### 2024-01-23 85f9465
Fix Mypy install in GH actions.
