### 2024-01-23 7a36330
Possible fix for Mypy config.
