+++ .github/workflows/mypy.yml
@@ -9,6 +9,11 @@ jobs:
     name: Mypy
     steps:
     - uses: actions/checkout@v4
+    
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: "3.11"
 
     - name: Install dependencies
       run: |
