### 2024-01-23 fd7469d
Specify path in Mypy action
