+++ .github/workflows/mypy.yml
@@ -23,5 +23,6 @@ jobs:
       with:
         github_token: ${{ secrets.github_token }}
         reporter: github-pr-review
+        execute_command: "mypy ."
         # The action will output fail if there are mypy errors
         level: error
