### 2024-01-24 40005b9
Fix Mypy error.
