+++ rdata/conversion/__init__.py
@@ -4,10 +4,15 @@ from ._conversion import (
     Converter as Converter,
     RBuiltin as RBuiltin,
     RBytecode as RBytecode,
+    REnvironment as REnvironment,
     RExpression as RExpression,
+    RExternalPointer as RExternalPointer,
     RFunction as RFunction,
     RLanguage as RLanguage,
     SimpleConverter as SimpleConverter,
+    SrcFile as SrcFile,
+    SrcFileCopy as SrcFileCopy,
+    SrcRef as SrcRef,
     convert as convert,
     convert_array as convert_array,
     convert_attrs as convert_attrs,
