+++ docs/index.rst
@@ -33,6 +33,7 @@ Its main advantages are:
    apilist
    auto_examples/index
    Try online! <https://rdata.readthedocs.io/en/latest/lite/lab/?path=auto_examples/plot_example.ipynb>
+   conversions
    contributors
 
 The package rdata is developed `on Github <http://github.com/vnmabus/rdata>`_.
