+++ .github/workflows/ruff.yml
@@ -1,8 +1,10 @@
 name: Ruff
-on: [push, pull_request]
+on: [push]
 jobs:
   ruff:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
       - uses: chartboost/ruff-action@v1
+        with:
+          args: check --output-format github
