+++ rdata/__init__.py
@@ -5,6 +5,7 @@ from importlib.resources import files
 from typing import TYPE_CHECKING, Final
 
 from . import conversion as conversion, parser as parser
+from ._read import read_rda as read_rda, read_rds as read_rds
 
 if TYPE_CHECKING:
     from .parser._parser import Traversable
