+++ rdata/tests/test_rdata.py
@@ -39,91 +39,77 @@ class SimpleTests(unittest.TestCase):
 
     def test_logical(self) -> None:
         """Test parsing of logical vectors."""
-        parsed = rdata.parser.parse_file(TESTDATA_PATH / "test_logical.rda")
-        converted = rdata.conversion.convert(parsed)
+        data = rdata.read_rda(TESTDATA_PATH / "test_logical.rda")
 
-        np.testing.assert_equal(converted, {
+        np.testing.assert_equal(data, {
             "test_logical": np.array([True, True, False, True, False]),
         })
 
     def test_nullable_logical(self) -> None:
         """Test parsing of logical vectors containing NA."""
-        parsed = rdata.parser.parse_file(
-            TESTDATA_PATH / "test_nullable_logical.rda",
-        )
-        converted = rdata.conversion.convert(parsed)
+        data = rdata.read_rda(TESTDATA_PATH / "test_nullable_logical.rda")
 
-        data = converted["test_nullable_logical"]
+        array = data["test_nullable_logical"]
         np.testing.assert_array_equal(
-            data.data,
+            array.data,
             np.array([True, False, True]),
         )
         np.testing.assert_array_equal(
-            data.mask,
+            array.mask,
             np.array([False, False, True]),
         )
 
     def test_nullable_int(self) -> None:
         """Test parsing of integer vectors containing NA."""
-        parsed = rdata.parser.parse_file(
-            TESTDATA_PATH / "test_nullable_int.rda",
-        )
-        converted = rdata.conversion.convert(parsed)
+        data = rdata.read_rda(TESTDATA_PATH / "test_nullable_int.rda")
 
-        data = converted["test_nullable_int"]
+        array = data["test_nullable_int"]
         np.testing.assert_array_equal(
-            data.data,
+            array.data,
             np.array([313, -12, -2**31]),
         )
         np.testing.assert_array_equal(
-            data.mask,
+            array.mask,
             np.array([False, False, True]),
         )
 
     def test_vector(self) -> None:
         """Test parsing of numerical vectors."""
-        parsed = rdata.parser.parse_file(TESTDATA_PATH / "test_vector.rda")
-        converted = rdata.conversion.convert(parsed)
+        data = rdata.read_rda(TESTDATA_PATH / "test_vector.rda")
 
-        np.testing.assert_equal(converted, {
+        np.testing.assert_equal(data, {
             "test_vector": np.array([1.0, 2.0, 3.0]),
         })
 
     def test_empty_string(self) -> None:
         """Test that the empty string is parsed correctly."""
-        parsed = rdata.parser.parse_file(TESTDATA_PATH / "test_empty_str.rda")
-        converted = rdata.conversion.convert(parsed)
+        data = rdata.read_rda(TESTDATA_PATH / "test_empty_str.rda")
 
-        np.testing.assert_equal(converted, {
+        np.testing.assert_equal(data, {
             "test_empty_str": [""],
         })
 
     def test_na_string(self) -> None:
         """Test that the NA string is parsed correctly."""
-        parsed = rdata.parser.parse_file(
-            TESTDATA_PATH / "test_na_string.rda",
-        )
-        converted = rdata.conversion.convert(parsed)
+        data = rdata.read_rda(TESTDATA_PATH / "test_na_string.rda")
 
-        np.testing.assert_equal(converted, {
+        np.testing.assert_equal(data, {
             "test_na_string": [None],
         })
 
     def test_complex(self) -> None:
         """Test that complex numbers can be parsed."""
[…]
