+++ rdata/_read.py
@@ -74,6 +74,7 @@ def read_rds(  # noqa: PLR0913
             .. code-block :: python
 
                 def constructor(obj, attrs):
+                    ...
 
             This dictionary can be used to support custom R classes. By
             default, the dictionary used is
@@ -152,6 +153,7 @@ def read_rda(  # noqa: PLR0913
             .. code-block :: python
 
                 def constructor(obj, attrs):
+                    ...
 
             This dictionary can be used to support custom R classes. By
             default, the dictionary used is
