+++ docs/simpleusage.rst
@@ -15,7 +15,7 @@ The common way of reading an R dataset is the following one:
     
 which results in
 
-.. code:: python
+.. code::
 
     {'test_vector': array([1., 2., 3.])}
 
@@ -93,7 +93,7 @@ Pandas :class:`~pandas.Categorical` objects:
     
 which has the following result:
 
-.. code:: python
+.. code::
 
     {'test_dataframe':   class  value
         1     b'a'      1
