+++ pyproject.toml
@@ -52,6 +52,7 @@ docs = [
   "myst-parser",
   "pydata-sphinx-theme",
   "sphinx>=3.1",
+  "sphinx-codeautolink",
   "sphinx-gallery",
 ]
 typing = [
