### 2024-01-25 b2cb362
Prevent Pydev to recurse in asv benchmarks.
