### 2024-01-25 #144 @vnmabus
So I am back, and I think that I implemented most of the requested changes (in the develop branch, not yet released). First, for @rich-iannone:

>     * [ ]  Badges for:
>
>       * [x]  Continuous integration and test coverage,
>       * [x]  Docs building (if you have a documentation website),
>       * [ ]  A [repostatus.org](https://www.repostatus.org/) badge,
>       * [ ]  Python versions supported,
>       * [x]  Current package version (on PyPI / Conda).

Missing badges have been added

> _NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the a badge for pyOpenSci peer-review will be provided upon acceptance.)_

>     * [ ]  Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file.

I am not sure if you reviewed the latest version, because I think the links were in the [examples documentation page](https://rdata.readthedocs.io/en/latest/auto_examples/index.html) before your review. Please, tell me if that is not enough.

>     * [ ]  The need for the package is clear

Again, I am not sure if you reviewed the latest version. I made changes to both the [README](https://github.com/vnmabus/rdata/blob/develop/README.rst) and the [main page of the documentation](https://rdata.readthedocs.io/) after submitting but before your review. Please tell me if something can be improved.

>     * [ ]  **Performance:** Any performance claims of the software been confirmed.

We do not have performance claims (although with the recent changes to the XDR parser, performance has been greatly improved).

>       * [ ]  Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

I added Ruff to the CI and fixed all violations.

>     * [ ]  **A short summary** describing the high-level functionality of the software
>
>     * [ ]  **Authors:** A list of authors with their affiliations
>
>     * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>
>     * [ ]  **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

I plan to write it after the review, if possible, to accommodate all the changes.

> The software is good and useful. I find myself having a need for this very thing. In terms of letting people know what the package does, it's somewhat terse at the moment (along the lines of 'if you know you know'). That said, it would be valuable to demonstrate (perhaps through examples or an end-to-end solution) the value of taking an R dataset in .rda/.rds format and incorporating it into a Python workflow. This could involve using some collection of R packages that don't have a good correspondance in Python, generating one or more .rds files, and finishing the work in Python (using Python packages that have no parallel in R, to drive home the point).
>
> Other use cases can be presented to get the casual user's imagination going, as well. This may involve hosting .rds files (maybe in a public GH repo) and demonstrating how to retrieve and convert the data. Another idea is to demonstrate how to extract datasets from R packages. I know that such demos would involve operations that exist outside of the functionality of the package but it can be inspiring to tie I/O related things to the use of the package itself. It sort of shows complete solutions to a problem that can involve your software in a specific, key step.

As mentioned before, I am not sure if you reviewed the latest changes because a set of [examples](https://rdata.readthedocs.io/en/latest/auto_examples/index.html) was added in response to @NickleDave comments. It is true that the examples do not illustrate a complete workflow, just the loading part. Do you consider that is really necessary?

>     * add some benchmarking information; compare with other solutions (could use https://github.com/airspeed-velocity/asv for this)

I added asv tests. Currently I only test the array parsing routines, as their performance has greatly improved recently, but if you have additional proposals for performance tests I will consider adding them.

>     * provide information on which R object translations are supported

I created a [page](https://rdata.readthedocs.io/en/latest/conversions.html) in the documentation for this. Please, tell me if this is what you had in mind.

>     * provide some examples on how different translated objects can be used in a real-world analysis

I am not sure if I follow you here. Could you please clarify your intention?

> _Package README and Documentation Site_
>
> The organization of the project's GH page is pretty good. Without a direct link, however, it's a bit hard to search for (appearing on the 4th page of results) and that's due to many similarly named projects. To circumvent this, maybe add a few more search tags in the About section of the project page. Also, include a link to the documentation in the About section (I know it's linked from a badge and within the README but having a link at the top makes a difference). It's great to see that the package is getting a fair amount of use (100+ usages); you can draw more attention to that by removing the empty `Packages` section in the sidebar (this'll bring that useful stat higher up on the page).

I added the link (thanks!) and removed the "Packages" section. I am unsure of which additional search tags I could use.

> A useful badge to have is the Python versions supported badge. The list of versions could match those that you're testing and the relevant `"Programming Language :: Python :: 3.*"` classifiers should be added to the `pyproject.toml` file.

Added.

> _Community Considerations_
>
> To make things better for community contributions, I suggest adding these two things:
>
>     * add an issue template
>
>     * add a pull request template

I added both of them.

> You already have a CONTRIBUTING document that describes ways to make contributions and this is great. The PR and issue templates can link to that document (this increases the chance that users will read it).

I linked to it in the PR template.

> I recommend adding a CODE_OF_CONDUCT.md file. These are very commonplace so you can get a file from another project's repo.

That was already added before your review IIRC.

> _Testing_
>
> It's good to see that testing is being done on a solid matrix of recent Python versions and platforms (Linux/Ubuntu, Windows, MacOS). There does need to be some updates to the actions used in the testing workflow file (`actions/checkout@v2` -> `actions/checkout@v4` and `actions/setup-python@v2` -> `actions/setup-python@v4`).

Done.

> Another recommendation is to run pylint for linting, perhaps through pre-commit (see https://pylint.pycqa.org/en/latest/user_guide/installation/pre-commit-integration.html).

I have finally chosen Ruff for its performance. I added it to CI in order to report violations (but I do not want them to be automatically fixed without human intervention, so it just checks them).

> Perhaps consider adding Python 3.12 to the testing matrix.

Added.

> If you want to go even further with testing, have different R versions on different platforms generate .rds/.rda files and test those directly.

I want to improve testing, and I had several ideas. I created [an issue](https://github.com/vnmabus/rdata/issues/35) for tracking that.

> _Citation File_
>
> The citation file should have a version field, updated every time a version of the software is released. This can be a bit of a pain but it makes for a better, more complete citation and tools exist to help with this (see https://github.com/citation-file-format/citation-file-format#tools-to-work-with-citationcff-files-wrench).

I am a bit unsure about it. Adding a version field has the risk that the user does a copy-paste of the generated Bibtex from the Github repo, which may not be the same version that he is using. If I do not add it, either the user cites the package or he retrieves the version from his copy if he is interested in reproducibility. In both cases, the information should not be wrong. Is there a compelling argument for having the version there?
