### 2024-01-25 #144 @vnmabus
Now, for @has2k1:

>       * [ ]  A [repostatus.org](https://www.repostatus.org/) badge,
>       * [ ]  Python versions supported,

Added.

>     * [ ]  Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file.

I just noted that the link is not in the README, added now.

>     * [ ]  If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.

I do not have it yet. I should probably include a comparison at least with pyreadr, but I am not sure how to frame it.

>     * [ ]  Citation information

There is a CITATION.cff file. I plan to add citation to the README after publishing in JOSS (if they approve it).

>     * [ ]  **Performance:** Any performance claims of the software been confirmed.

I made no performance claims.

>     * [ ]  **Continuous Integration:** Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review)

I have tests, typing and style as Github actions.

>       * [ ]  Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

I added Ruff as a linter and fixed all violations.

>     * [ ]  **A short summary** describing the high-level functionality of the software
>
>     * [ ]  **Authors:** A list of authors with their affiliations
>
>     * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>
>     * [ ]  **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

I plan to add it after the review by PyOpenSci is complete.

>     1. For the code in the README (and since as it is not a doctest), it would be convenient to be able to copy & paste it and run without any editing.

Done.

>     2. I feel like there is are missing `read` functions that wraps `rdata.parser.parse_file` and
>        `rdata.parser.convert` and works with just a filepath. e.g.
>
>
> ```python
> parsed = rdata.parser.parse_file("dataframe.rds")
> data = rdata.conversion.convert(parsed)
> ```
>
> would be
>
> ```python
> data = rdata.read_rds("dataframe.rds")
> ```
>
> And maybe also a `read_rda`. A single `read_rdata` can do for both cases but that does not communicate the difference between reading an object vs an environment with objects.
>
> These functions could include optional arguments to pass on to the _parser_ and/or _converter_.

These convenience functions have been incorporated.

>     3. It is not documented anywhere what base R objects are supported and what they are translate to as python object. I think a table would be helpful.

I added a [page in the documentation](https://rdata.readthedocs.io/en/latest/conversions.html). Please tell me if that is what you had in mind.

>     4. On packaging configuration, there is a `setup.cfg` and `pyproject.toml`. As even the configuration for `setuptools` is in `pyproject.toml`, it makes sense to migrate the configs for `isort`, `mypy` and `pytest`. That would leave only the `flake8` configuration in `setup.cfg` as `flake8` [does not yet support](https://github.com/PyCQA/flake8/issues/234) `pyproject.toml`.

Everything moved to `pyproject.toml`. Flake8 config has been removed as now the project uses Ruff.

>     5. `xdrlib` has been deprecated in python 3.12 and it is good there is a plan (and some work) to deal with it.

The PR has been merged and `xdrlib` is no longer required.
