### 2024-01-26 d43adb7
Add xdr and ascii writers
