+++ rdata/io/base.py
@@ -1,0 +1,192 @@
+from __future__ import annotations
+
+import abc
+import warnings
+
+import numpy as np
+
+from rdata.parser._parser import (
+        RObjectType,
+        RObjectInfo,
+        RVersions,
+        RExtraInfo,
+        )
+
+
+def pack_r_object_info(info: RObjectInfo) -> np.int32:
+    if info.type == RObjectType.NILVALUE:
+        bits = f'{0:24b}'
+    elif info.type == RObjectType.REF:
+        bits = f'{info.reference:24b}'
+    else:
+        bits = (f'{0:4b}'
+                f'{info.gp:16b}'
+                f'{0:1b}'
+                f'{info.tag:1b}'
+                f'{info.attributes:1b}'
+                f'{info.object:1b}'
+                )
+    bits += f'{info.type.value:8b}'
+    bits = bits.replace(' ', '0')
+    assert len(bits) == 32
+    info_int = np.packbits([int(b) for b in bits]).view('>i4').astype('=i4')[0]
+    return info_int
+
+
+class Writer(abc.ABC):
+    """Writer interface for a R file."""
+
+    def __init__(self):
+        pass
+
+    @abc.abstractmethod
+    def write_magic(self):
+        pass
+
+    def write_header(self, versions: RVersions, extra: RExtraInfo):
+        """Write header."""
+        self.write_int(versions.format)
+        self.write_int(versions.serialized)
+        self.write_int(versions.minimum)
+        if versions.format >= 3:
+            self.write_string(extra.encoding.encode('ascii'))
+
+    def write_bool(self, value):
+        self.write_int(int(value))
+
+    @abc.abstractmethod
+    def write_nullable_bool(self, value):
+        pass
+
+    def write_int(self, value):
+        if not isinstance(value, (int, np.int32)):
+            raise RuntimeError(f'Not valid integer: {value} ({type(value)})')
+        self.write_nullable_int(value)
+
+    @abc.abstractmethod
+    def write_nullable_int(self, value):
+        pass
+
+    @abc.abstractmethod
+    def write_double(self, value):
+        pass
+
+    def write_complex(self, value):
+        self.write_double(value.real)
+        self.write_double(value.imag)
+
+    @abc.abstractmethod
+    def write_string(self, value):
+        pass
+
+    def __write_array(self, array, write_value):
+        array = np.ravel(array)
+        self.write_int(len(array))
+        # XXX check if row or column major
+        for value in array:
+            write_value(value)
+
+    def write_nullable_bool_array(self, array):
+        """Write a boolean array."""
+        self.__write_array(array, self.write_nullable_bool)
+
+    def write_nullable_int_array(self, array):
+        """Write an integer array."""
+        self.__write_array(array, self.write_nullable_int)
+
+    def write_double_array(self, array):
+        """Write a double array."""
+        self.__write_array(array, self.write_double)
+
+    def write_complex_array(self, array):
+        """Write a complex array."""
+        self.__write_array(array, self.write_complex)
+
+    def write_all(self, data):
+        self.write_magic()
+        self.write_header(data.versions, data.extra)
+        self.write_R_object(data.object)
+
+    def write_R_object(self, obj):
+        # Some types write attributes and tag with data while some write them
+        # later. These booleans keep track of whether attributes or tag
+        # has been written already
+        attributes_written = False
+        tag_written = False
+
+        # Write info bytes
+        info = obj.info
+        self.write_int(pack_r_object_info(info))
+
+        # Write data
+        value = obj.value
+        if info.type in {
+           RObjectType.NIL,
+           RObjectType.NILVALUE,
+        }:
+            # These types don't have any data
+            pass
+
+        elif info.type == RObjectType.SYM:
+            self.write_R_object(value)
+
+        elif info.type in {
+            RObjectType.LIST,
[…]
