### 2024-01-26 2331898
Include non-zero gp value in representation
