### 2024-02-14 vnmabus/rdata#38 @has2k1
How does this package compare to pyreadr?
### Motivation

How does this package compare to pyreadr?

### Desired functionality

.

### Alternatives

_No response_

### Additional context

_No response_
