### 2024-02-14 #144 @has2k1
@vnmabus has resolved the all core queries and addressed the 2 conditional issues.

To keep track of one, I have filed an [issue](https://github.com/vnmabus/rdata/issues/38) about the comparison with other python packages, and the `paper.md` file for JOSS submission can be resolved when it is time to submit.

In between, the tables for default [conversions](https://rdata.readthedocs.io/en/latest/conversions.html) are great and super informative.

Great work Carlos.
