### 2024-02-14 vnmabus/rdata#38 @vnmabus
As far as I understand, pyreadr is mostly a wrapper around [librdata](https://github.com/WizardMac/librdata). Thus, it have the same limitations:
- pyreadr is not a pure Python package, with all the complications that it causes for the installation. On the other hand, rdata is a pure Python package, and this allows for example using it in the browser via Pyodide.
- Due to it being a wrapper around C code, the development of pyreadr is slower, as they must wait for the underlying C library to update (see https://github.com/ofajardo/pyreadr/issues?q=label%3A%22waiting+for+librdata+changes%22+), while the rdata package is pure Python and thus easier to understand and modify.
- As a consequence, pyread is not able to read many types that rdata can. See for example https://github.com/ofajardo/pyreadr/issues/8 and https://github.com/ofajardo/pyreadr/issues/51.
- The license for pyread is one of the least permissive licenses in the open-source world, the [GNU Affero General Public License](https://www.gnu.org/licenses/agpl-3.0.en.html). That means that you cannot have it as a dependence in a normal Python library, which use often MIT or BSD licenses, without changing the license. rdata is MIT instead, so it does not have that problem.

I want also to say that rdata predates pyreadr for almost a year, although the latter has done a bigger effort to attract users (I did not have time to write good documentation until one or two years ago).
