### 2024-02-29 295d29c
Add initial test for writing
