### 2024-02-29 c975802
Write RDA magic with other magic
