### 2024-02-29 9453eee
Include test for ascii files
