+++ rdata/tests/test_write.py
@@ -28,15 +28,21 @@ def decompress_data(data):
 @pytest.mark.parametrize("fname", [
     "test_vector.rda",
     "test_full_named_matrix.rds",
+    "test_ascii_v2.rda",
     ])
 def test_write(fname):
     with (TESTDATA_PATH / fname).open("rb") as f:
         data = decompress_data(f.read())
         rds = data[:2] != b'RD'
+        format = 'ascii' if data.isascii() else 'xdr'
 
         r_data = rdata.parser.parse_data(data)
 
-        fd = io.BytesIO()
-        rdata.io.write_file(fd, r_data, format='xdr', rds=rds)
+        if format == 'ascii':
+            fd = io.StringIO()
+            data = data.decode('ascii')
+        else:
+            fd = io.BytesIO()
+        rdata.io.write_file(fd, r_data, format=format, rds=rds)
 
         assert data == fd.getvalue()
