### 2024-02-29 192362c
Do not include trailing .0 for doubles
