### 2024-02-29 98c29b0
Loop over all test files
