### 2024-02-29 cb15e4c
Fix testing Windows-generated ascii files
