### 2024-02-29 646b32d
Fix writing NA string
