### 2024-02-29 ab5915a
Add debugging output
