### 2024-02-29 ebef7d1
Remove redundant writer functions
