+++ rdata/io/xdr.py
@@ -9,15 +9,6 @@ from rdata.parser._parser import R_INT_NA
 from .base import Writer
 
 
-def flatten_nullable_int_array(array):
-    assert array.dtype in (np.int32, int)
-    if np.ma.is_masked(array):
-        mask = np.ma.getmask(array)
-        array = np.ma.getdata(array).copy()
-        array[mask] = R_INT_NA
-    return array
-
-
 class WriterXDR(Writer):
     """Writer for files in XDR format."""
 
@@ -32,15 +23,23 @@ class WriterXDR(Writer):
             self.file.write(f"RDX{rda_version}\n".encode("ascii"))
         self.file.write(b"X\n")
 
-    def __write_array(self, array):
-        # Expect only 1D arrays here
-        assert array.ndim == 1
-        self.write_int(array.size)
-        self.__write_array_values(array)
+    def _write_array_values(self, array):
+        # Convert boolean to int
+        if np.issubdtype(array.dtype, np.bool_):
+            array = array.astype(np.int32)
 
-    def __write_array_values(self, array):
+        # Convert int arrays to int32 and flatten masked values
+        if np.issubdtype(array.dtype, np.integer):
+            array = array.astype(np.int32)
+            if np.ma.is_masked(array):
+                mask = np.ma.getmask(array)
+                array = np.ma.getdata(array).copy()
+                array[mask] = R_INT_NA
+
         # Convert to big endian if needed
         array = array.astype(array.dtype.newbyteorder(">"))
+
+        # Create a contiguous data buffer if not already
         # 1D array should be both C and F contiguous
         assert array.flags["C_CONTIGUOUS"] == array.flags["F_CONTIGUOUS"]
         if array.flags["C_CONTIGUOUS"]:
@@ -48,35 +47,6 @@ class WriterXDR(Writer):
         else:
             data = array.tobytes()
         self.file.write(data)
-
-    def write_nullable_bool(self, value):
-        if value is None or np.ma.is_masked(value):
-            value = R_INT_NA
-        self.__write_array_values(np.array(value).astype(np.int32))
-
-    def write_nullable_int(self, value):
-        if value is None or np.ma.is_masked(value):
-            value = R_INT_NA
-        self.__write_array_values(np.array(value).astype(np.int32))
-
-    def write_double(self, value):
-        self.__write_array_values(np.array(value))
-
-    def write_complex(self, value):
-        self.__write_array_values(np.array(value))
-
-    def write_nullable_bool_array(self, array):
-        self.write_nullable_int_array(array.astype(np.int32))
-
-    def write_nullable_int_array(self, array):
-        array = flatten_nullable_int_array(array)
-        self.__write_array(array.astype(np.int32))
-
-    def write_double_array(self, array):
-        self.__write_array(array)
-
-    def write_complex_array(self, array):
-        self.__write_array(array)
 
     def write_string(self, value: bytes):
         if value is None:
