### 2024-02-29 cd30d46
Fix empty (None) string
