### 2024-02-29 b153a3d
Add test for conversion
