### 2024-02-29 5a9994b
Add option to change version information
