### 2024-02-29 dea4548
Set gp flag as plain int
