### 2024-02-29 4bace9f
Fix extra info in version 2 format
