### 2024-02-29 65d1468
Generalize list builder
