### 2024-02-29 b97d650
Use same encoding as in input
