### 2024-03-01 #144 @isabelizimm
With two reviewers giving the 👍 ... it is my absolute pleasure to say that ... 🎉 ... `rdata` has been approved by pyOpenSci!

Thank you @vnmabus for submitting `rdata` and many thanks to @has2k1 and @rich-iannone for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of `rdata`. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] task for @isabelizimm (that's me!) Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉

---
