### 2024-03-01 d99a003
Add equality operator for comparing numpy arrays
