### 2024-03-01 0ae79e0
Print also python data for debugging
