### 2024-03-01 fb48601
Remove unused variable
