### 2024-03-04 534fa47
Allow debug printing in tests
