### 2024-03-04 b82aa04
Check that integers are small enough
