### 2024-03-04 8aec705
Simplify loops
