### 2024-03-04 da0b8c9
Add PyOpenSci badge.
