+++ rdata/__init__.py
@@ -21,4 +21,4 @@ Path of the test data.
 
 """
 
-__version__ = "0.10.1.dev0"
+__version__ = "0.11.0.dev0"
