+++ rdata/__init__.py
@@ -21,4 +21,4 @@ Path of the test data.
 
 """
 
-__version__ = "0.11.2"
+__version__ = "0.11.3.dev0"
