### 2024-03-07 a16f99a
Execute Ruff in PR CI.
