+++ .github/workflows/ruff.yml
@@ -1,5 +1,5 @@
 name: Ruff
-on: [push]
+on: [push, pull_request]
 jobs:
   ruff:
     runs-on: ubuntu-latest
