### 2024-03-12 #144 @lwasser
hey @vnmabus just a quick note that i saw this message and noticed a bug in our website! things should be working and rdata should filter as expected now:
thanks for letting us know about this 🙌 and welcome to our community!

![Image](https://github.com/pyOpenSci/software-submission/assets/7649194/630b0a91-e027-47db-b922-d86b0fe071df)
