### 2024-03-26 c0a118e
Remove debug prints from tests
