### 2024-03-27 e0d5e6d
Add test for writing files with compression
