### 2024-03-27 9bba69b
Use binary mode for writing ascii files
