### 2024-04-15 386f907
Do not open temporary file twice
