### 2024-05-06 fbdbbe0
Add high-level writer functions
