### 2024-05-06 ad5ffc9
Add test for bad RDA creation
