### 2024-05-06 5c36c73
Add tests for bad encodings
