### 2024-05-06 b2b9b2b
Include testing None values
