### 2024-05-06 5c27b04
Raise NotImplementedError on untested code
