### 2024-05-06 df5b463
Remove unused function
