### 2024-05-06 4beb10c
Add test for writing too large integers
