### 2024-05-06 5e51d1f
Raise error instead of warning for unwritten tag
