### 2024-05-06 3d815f1
Exclude type checking blocks from coverage report
