+++ pyproject.toml
@@ -103,6 +103,11 @@ addopts = "--doctest-modules --doctest-glob='*.rst'"
 doctest_optionflags = "NORMALIZE_WHITESPACE ELLIPSIS"
 norecursedirs = ".* build dist *.egg venv .svn _build docs/auto_examples examples asv_benchmarks"
 
+[tool.coverage.report]
+exclude_also = [
+    "if TYPE_CHECKING:",
+    ]
+
 [tool.ruff.lint]
 select = [
     "ALL",
