+++ rdata/unparser/_unparser.py
@@ -1,9 +1,8 @@
-"""Abstract base class for writers."""
+"""Abstract base class for unparsers."""
 
 from __future__ import annotations
 
 import abc
-import warnings
 from typing import TYPE_CHECKING, Any
 
 import numpy as np
@@ -41,61 +40,61 @@ def pack_r_object_info(info: RObjectInfo) -> np.int32:
     return np.packbits([int(b) for b in bits]).view(">i4").astype("=i4")[0]  # type: ignore [no-any-return]
 
 
-class Writer(abc.ABC):
-    """Writer interface for a R file."""
+class Unparser(abc.ABC):
+    """Unparser interface for a R file."""
 
     @abc.abstractmethod
-    def write_magic(self, rda_version: int | None) -> None:
-        """Write magic bits."""
+    def unparse_magic(self, rda_version: int | None) -> None:
+        """Unparse magic bits."""
 
-    def write_header(self, versions: RVersions, extra: RExtraInfo) -> None:
-        """Write header."""
-        self.write_int(versions.format)
-        self.write_int(versions.serialized)
-        self.write_int(versions.minimum)
+    def unparse_header(self, versions: RVersions, extra: RExtraInfo) -> None:
+        """Unparse header."""
+        self.unparse_int(versions.format)
+        self.unparse_int(versions.serialized)
+        self.unparse_int(versions.minimum)
         minimum_version_with_encoding = 3
         if versions.format >= minimum_version_with_encoding:
             assert extra.encoding is not None
-            self.write_string(extra.encoding.encode("ascii"))
+            self.unparse_string(extra.encoding.encode("ascii"))
 
-    def write_int(self, value: int | np.int32) -> None:
-        """Write an integer value."""
-        self._write_array_values(np.array([value]))
+    def unparse_int(self, value: int | np.int32) -> None:
+        """Unparse an integer value."""
+        self._unparse_array_values(np.array([value]))
 
-    def write_array(self, array: npt.NDArray[Any]) -> None:
-        """Write an array of values."""
+    def unparse_array(self, array: npt.NDArray[Any]) -> None:
+        """Unparse an array of values."""
         # Expect only 1D arrays here
         assert array.ndim == 1
-        self.write_int(array.size)
-        self._write_array_values(array)
+        self.unparse_int(array.size)
+        self._unparse_array_values(array)
 
     @abc.abstractmethod
-    def _write_array_values(self, array: npt.NDArray[Any]) -> None:
-        """Write the values of an array."""
+    def _unparse_array_values(self, array: npt.NDArray[Any]) -> None:
+        """Unparse the values of an array."""
 
     @abc.abstractmethod
-    def write_string(self, value: bytes) -> None:
-        """Write a string."""
+    def unparse_string(self, value: bytes) -> None:
+        """Unparse a string."""
 
-    def write_r_data(self, r_data: RData, *, rds: bool = True) -> None:
-        """Write an RData object."""
-        self.write_magic(None if rds else r_data.versions.format)
-        self.write_header(r_data.versions, r_data.extra)
-        self.write_r_object(r_data.object)
+    def unparse_r_data(self, r_data: RData, *, rds: bool = True) -> None:
+        """Unparse an RData object."""
+        self.unparse_magic(None if rds else r_data.versions.format)
+        self.unparse_header(r_data.versions, r_data.extra)
+        self.unparse_r_object(r_data.object)
 
-    def write_r_object(self, obj: RObject) -> None:  # noqa: C901, PLR0912
-        """Write an RObject object."""
-        # Some types write attributes and tag with data while some write them
+    def unparse_r_object(self, obj: RObject) -> None:  # noqa: C901, PLR0912
+        """Unparse an RObject object."""
[…]

[…]
