### 2024-05-16 1c1cbfb
Add unparse_data() function returning bytes
