### 2024-05-16 c129729
Fix function names in docstring
