### 2024-06-04 6d504d7
Use Literal type for compression
