### 2024-06-04 07f4393
Use None for no compression
