### 2024-06-04 f94996a
Use Literal type for file type
