### 2024-06-04 746848d
Simplify handling RDA magic
