### 2024-06-05 d834a8a
Simplify handling RDA vs RDS difference
