### 2024-06-05 35c7ba8
Simplify type names
