### 2024-06-05 3b201d2
Add test for unparsing bad rda
