### 2024-06-06 86958d6
Simplify handling version integers in write
