### 2024-06-06 49f70a1
Use hex representation for R versions
