### 2024-06-06 7509ecb
Use callback protocol for type hint
