### 2024-06-06 7eb953b
Fix mypy in Python 3.12
