### 2024-06-06 9b3093c
Remove unnecessary casting
