### 2024-06-06 d82995f
Fix converting empty dict
