### 2024-06-06 1fef633
Use file type parser function
