### 2024-06-06 08c513f
Add tests for parsing empty lists
