### 2024-06-06 0e001ea
Allow only non-empty data
