### 2024-06-06 c35c0e5
Add test for failing to create empty RDA file
