### 2024-06-06 f84b4aa
Include received type in the error message
