### 2024-06-06 7aa1d31
Clarify the conversion of np.array([None])
