### 2024-06-06 38c1ff6
Simplify no-error context
