### 2024-06-06 cf78843
Rewrite file checker with existing functions
