### 2024-06-06 d4c961f
Clarify compression handling
