### 2024-06-06 4f45cba
Add test for parsing NaN and Inf
