### 2024-06-06 2bc73ce
Fix comparing NaN values
