### 2024-06-06 5ab4cd5
Fix writing NaN and Inf
