### 2024-06-06 68b4e55
Simplify packing bits
