### 2024-06-11 #144 @vnmabus
Sorry, I got a new job the same day this package was accepted in PyOpenSci, so I had a few months without a lot of time to invest on it. My plan is still to submit to JOSS after https://github.com/vnmabus/rdata/pull/40 is merged.
