### 2024-06-12 #144 @lwasser
hey @vnmabus no worries. no need to apologize!

A few notes

1. because JOSS accepts the review that we implemented and as such does NOT re-review your code, adding code to your package via tha referenced pr which looks significant between acceptance here and publication at JOSS is not ideal. The release for this review - 0.11.0 should be the same accepted by JOSS. they will ask you to create a new release with the joss badge and doi i believe.
2. the review from joss should be a fast track - they will only review your paper - not the code. so it should be quick and not a lot of additional effort on your part (aside from writing the paper!).

there is no huge rush but I highly encourage you to submit to joss before adding more code to your package to keep our existing review synced with what joss accepts as a publication.
If not, there will be an issue with a lot of new functionality that has not been reviewed associated with a doi.

please let me know if you have any questions!
