### 2024-06-13 #144 @lwasser
@vnmabus i'm going to ping @arfon on this specific issue from JOSS. We ask authors to submit a package to us when the API has stablized and similar to an academic paper being reviewed, it is not ideal if changes are being made while the software / (or paper!) is in active review.

Arfon - in this case we have reviewed the rdata package and it has been pyOpenSci accepted. However, since the pyOpenSci acceptance of rdata there are new commits that look to me like fairly significant enhancements that have been added to the package.

Because there is new functionality, we may not want to move forward with a fast track (these changes might require another review). Generally we ask the maintainer to submit based on the release / doi created at the end of our pyOpenSci review and as such JOSS can trust that we have reviewed the entire code base. In this case we have a divergence.

How would you prefer to handle this on the JOSS side of things? I believe that this is the first time this has happened. many thanks!
