### 2024-06-17 42ac9b8
Use Literal type and lower case for encodings
