### 2024-06-17 7cf643f
Fix integer range check in NumPy 2.0
