### 2024-06-22 #144 @arfon
Hrm, yes isn't ideal (I agree with the idea that the software should essentially be the same between the pyOpenSci and JOSS reviews). That said, we do have the ability to associate a specific version of a software package with a JOSS publication (e.g., `v1.0.0`) so provided there is a release/version number that pertains to the pyOpenSci submission, we could simply do that?
