### 2024-08-25 #144 @arfon
@vnmabus – you're welcome to include future-looking work in the paper, and @trossi as author too given their current role on the project.

> @arfon would it make sense for @vnmabus to submit the package to review for JOSS noting this specific item (when they have bandwidth!)? And you / JOSS can make a call on addressing the new features in that paper over there? We will link back to the review here as we always do so there is a clear communication trail.

Sounds good to me!
