+++ paper/paper.bib
@@ -1,0 +1,53 @@
+@misc{diaz-vico+ramos-carreno_2022_scikitdatasets,
+  title = {{{scikit-datasets}}: {{Scikit-learn-compatible}} Datasets},
+  author = {{D{\'i}az-Vico}, David and {Ramos-Carre{\~n}o}, Carlos},
+  year = {2022},
+  month = mar,
+  doi = {10.5281/zenodo.6383047},
+  url = {https://github.com/daviddiazvico/scikit-datasets},
+  copyright = {MIT}
+}
+
+@misc{fajardo_2018_pyreadr,
+  title = {Pyreadr},
+  author = {Fajardo, Otto},
+  year = {2018},
+  month = dec,
+  doi = {10.5281/zenodo.7110170},
+  url = {https://github.com/ofajardo/pyreadr}
+}
+
+@misc{gautier_2024_rpy2,
+  title = {Rpy2: {{R}} in {{Python}}},
+  author = {Gautier, Laurent},
+  year = {2024},
+  publisher = {GitHub},
+  url = {https://github.com/rpy2/rpy2}
+}
+
+@software{pandasdevelopmentteam_2020_pandasdev,
+  title = {{{pandas-dev/pandas}}: {{pandas}}},
+  author = {Pandas Development Team},
+  year = {2020},
+  month = feb,
+  publisher = {Zenodo},
+  doi = {10.5281/zenodo.3509134},
+  url = {https://doi.org/10.5281/zenodo.3509134},
+  version = {latest}
+}
+
+@article{ramos-carreno+_2024_scikitfda,
+  title = {Scikit-Fda: {{A Python Package}} for {{Functional Data Analysis}}},
+  shorttitle = {Scikit-Fda},
+  author = {{Ramos-Carre{\~n}o}, Carlos and Torrecilla, Jos{\'e} Luis and {Carbajo-Berrocal}, Miguel and Marcos, Pablo and Su{\'a}rez, Alberto},
+  year = {2024},
+  month = may,
+  journal = {Journal of Statistical Software},
+  volume = {109},
+  pages = {1--37},
+  issn = {1548-7660},
+  doi = {10.18637/jss.v109.i02},
+  abstract = {The library scikit-fda is a Python package for functional data analysis (FDA). It provides a comprehensive set of tools for representation, preprocessing, and exploratory analysis of functional data. The library is built upon and integrated in Python's scientific ecosystem. In particular, it conforms to the scikit-learn application programming interface so as to take advantage of the functionality for machine learning provided by this package: Pipelines, model selection, and hyperparameter tuning, among others. The scikit-fda package has been released as free and open-source software under a 3-clause BSD license and is open to contributions from the FDA community. The library's extensive documentation includes step-by-step tutorials and detailed examples of use.},
+  copyright = {Copyright (c) 2024 Carlos Ramos-Carre{\~n}o, Jos{\'e} Luis Torrecilla, Miguel Carbajo-Berrocal, Pablo Marcos, Alberto Su{\'a}rez},
+  langid = {english}
+}
