### 2024-09-05 4557559
Add helper function for creating unicode arrays
