### 2024-09-05 6eeb992
Add more pd.Series types
