### 2024-09-07 e074b3f
Apply suggestions from code review
Co-authored-by: Tuomas Rossi <34502776+trossi@users.noreply.github.com>
