### 2024-09-09 6c0849a
Add test for all ascii characters
