### 2024-09-09 26a626c
Fix escaping various characters for R output
