### 2024-09-10 eb82ff6
Add a converter class for Python-to-R conversion
