+++ rdata/conversion/to_r.py
@@ -91,7 +91,7 @@ def build_r_object(
         attributes: RObject | None = None,
         tag: RObject | None = None,
         gp: int = 0,
-        reference: int = 0,
+        reference: tuple(int, RObject | None) = (0, None),
 ) -> RObject:
     """
     Build R object.
@@ -102,7 +102,7 @@ def build_r_object(
         attributes: Same as in RObject.
         tag: Same as in RObject.
         gp: Same as in RObjectInfo.
-        reference: Same as in RObjectInfo.
+        reference: Tuple of integer and object.
 
     Returns:
         R object.
@@ -112,10 +112,8 @@ def build_r_object(
         RObjectInfo
     """
     assert r_type is not None
-    if reference == 0:
-        assert reference_name_list[reference] is None
-    else:
-        assert r_type == RObjectType.REF
+    reference_id, referenced_object = reference
+    assert (reference_id == 0) == (referenced_object == None) == (r_type != RObjectType.REF)
     is_object = find_is_object(attributes)
     return RObject(
         RObjectInfo(
@@ -124,96 +122,13 @@ def build_r_object(
             attributes=attributes is not None,
             tag=tag is not None,
             gp=gp,
-            reference=reference,
+            reference=reference_id,
          ),
          value,
          attributes,
          tag,
-         reference_obj_list[reference],
+         referenced_object,
      )
-
-
-def build_r_list(
-        data: Mapping[str, Any] | list[Any],
-        *,
-        encoding: Encoding,
-        convert_value: Converter | None = None,
-) -> RObject:
-    """
-    Build R object representing named linked list.
-
-    Args:
-        data: Non-empty dictionary or list.
-        encoding: Encoding to be used for strings within data.
-        convert_value: Function used for converting value to R object
-            (for example, convert_to_r_object).
-
-    Returns:
-        R object.
-    """
-    if convert_value is None:
-        convert_value = convert_to_r_object
-
-    if len(data) == 0:
-        msg = "data must not be empty"
-        raise ValueError(msg)
-
-    if isinstance(data, dict):
-        data = data.copy()
-        key = next(iter(data))
-        tag = build_r_sym(key, encoding=encoding)
-        value1 = convert_value(data.pop(key), encoding=encoding)
-    elif isinstance(data, list):
-        value1 = convert_value(data[0], encoding=encoding)
-        data = data[1:]
-        tag = None
-
-    if len(data) == 0:
-        value2 = build_r_object(RObjectType.NILVALUE)
-    else:
-        value2 = build_r_list(data, encoding=encoding, convert_value=convert_value)
-
-    return build_r_object(
-        RObjectType.LIST,
-        value=(value1, value2),
-        tag=tag,
-        )
-
-
-# XXX global lists
-reference_name_list = [None]
-reference_obj_list = [None]
-
-
-def build_r_sym(
-        name: str,
-        *,
-        encoding: Encoding,
-) -> RObject:
-    """
-    Build R object representing symbol.
-
-    Args:
-        name: String.
-        encoding: Encoding to be used for the name.
-
-    Returns:
-        R object.
-    """
-    # Reference to existing symbol if exists
-    if name in reference_name_list:
-        # XXX can any symbol be referenced???
-        reference = reference_name_list.index(name)
-        return build_r_object(RObjectType.REF, reference=reference)
-
-    # Create a new symbol
-    r_value = convert_to_r_object(name.encode(encoding), encoding=encoding)
-    r_object = build_r_object(RObjectType.SYM, value=r_value)
-
-    # Add to reference list
-    reference_name_list.append(name)
-    reference_obj_list.append(r_object)
-    return r_object
 
 
 def build_r_data(
[…]
