+++ rdata/tests/test_write.py
@@ -10,6 +10,7 @@ from typing import TYPE_CHECKING, Any
 import pytest
 
 import rdata
+from rdata.conversion import ConverterFromPythonToR, build_r_data
 from rdata.unparser import unparse_data
 
 if TYPE_CHECKING:
@@ -127,13 +128,12 @@ def test_convert_to_r(fname: str) -> None:
             encoding = encoding.lower()  # type: ignore [assignment]
 
         try:
+            converter = ConverterFromPythonToR(encoding=encoding)
             if file_type == "rds":
-                r_obj = rdata.conversion.convert_to_r_object(
-                    py_data, encoding=encoding)
+                r_obj = converter.convert_to_r_object(py_data)
             else:
-                r_obj = rdata.conversion.convert_to_r_object_for_rda(
-                    py_data, encoding=encoding)
-            new_r_data = rdata.conversion.build_r_data(
+                r_obj = converter.convert_to_r_object_for_rda(py_data)
+            new_r_data = build_r_data(
                 r_obj,
                 encoding=encoding,
                 format_version=r_data.versions.format,
@@ -150,21 +150,24 @@ def test_convert_to_r_bad_rda() -> None:
     """Test checking that data for RDA has variable names."""
     py_data = "hello"
     with pytest.raises(TypeError, match="(?i)data must be a dictionary"):
-        rdata.conversion.convert_to_r_object_for_rda(py_data)  # type: ignore [arg-type]
+        converter = ConverterFromPythonToR()
+        converter.convert_to_r_object_for_rda(py_data)  # type: ignore [arg-type]
 
 
 def test_convert_to_r_empty_rda() -> None:
     """Test checking that data for RDA has variable names."""
     py_data: dict[str, Any] = {}
     with pytest.raises(ValueError, match="(?i)data must not be empty"):
-        rdata.conversion.convert_to_r_object_for_rda(py_data)
+        converter = ConverterFromPythonToR()
+        converter.convert_to_r_object_for_rda(py_data)
 
 
 def test_unparse_bad_rda() -> None:
     """Test checking that data for RDA has variable names."""
     py_data = "hello"
-    r_obj = rdata.conversion.convert_to_r_object(py_data)
-    r_data = rdata.conversion.build_r_data(r_obj)
+    converter = ConverterFromPythonToR()
+    r_obj = converter.convert_to_r_object(py_data)
+    r_data = build_r_data(r_obj)
     with pytest.raises(ValueError, match="(?i)must be dictionary-like"):
         unparse_data(r_data, file_type="rda")
 
@@ -172,20 +175,23 @@ def test_unparse_bad_rda() -> None:
 def test_convert_to_r_bad_encoding() -> None:
     """Test checking encoding."""
     with pytest.raises(LookupError, match="(?i)unknown encoding"):
-        rdata.conversion.convert_to_r_object("ä", encoding="non-existent")  # type: ignore [arg-type]
+        converter = ConverterFromPythonToR(encoding="non-existent")
+        converter.convert_to_r_object("ä")  # type: ignore [arg-type]
 
 
 def test_convert_to_r_unsupported_encoding() -> None:
     """Test checking encoding."""
     with pytest.raises(ValueError, match="(?i)unsupported encoding"):
-        rdata.conversion.convert_to_r_object("ä", encoding="cp1250")  # type: ignore [arg-type]
+        converter = ConverterFromPythonToR(encoding="cp1250")
+        converter.convert_to_r_object("ä")  # type: ignore [arg-type]
 
 
 def test_unparse_big_int() -> None:
     """Test checking too large integers."""
     big_int = 2**32
[…]
