### 2024-09-10 1868d8a
Fix masked values in masked array
