### 2024-09-10 8d9cb55
Compare first string representations
