### 2024-09-10 9cdd37c
Add support for dataframe with string index
