### 2024-09-12 5084d2d
Add assertions for strings
