### 2024-09-12 af0f6fe
Add conversion for rangeindex and range
