### 2024-09-12 b205d8d
Move build_r_data function under converter class
