+++ rdata/_write.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from .conversion import build_r_data, ConverterFromPythonToR
+from .conversion import ConverterFromPythonToR
 from .conversion.to_r import DEFAULT_FORMAT_VERSION
 from .unparser import unparse_file
 
@@ -27,10 +27,7 @@ def write_rds(
     """
     Write an RDS file.
 
-    This is a convenience function that wraps
-    :func:`rdata.conversion.convert_to_r_object`,
-    :func:`rdata.conversion.build_r_data`,
-    and :func:`rdata.unparser.unparse_file`,
+    This is a convenience function that wraps conversion and unparsing
     as it is the common use case.
 
     Args:
@@ -52,12 +49,13 @@ def write_rds(
         >>> data = ["hello", 1, 2.2, 3.3+4.4j]
         >>> rdata.write_rds("test.rds", data)
     """
-    r_object = ConverterFromPythonToR(encoding=encoding).convert_to_r_object(data)
-    r_data = build_r_data(
-        r_object,
+    converter = ConverterFromPythonToR(
         encoding=encoding,
         format_version=format_version,
     )
+    r_object = converter.convert_to_r_object(data)
+    r_data = converter.build_r_data(r_object)
+
     unparse_file(
         path,
         r_data,
@@ -79,10 +77,7 @@ def write_rda(
     """
     Write an RDA or RDATA file.
 
-    This is a convenience function that wraps
-    :func:`rdata.conversion.convert_to_r_object_for_rda`,
-    :func:`rdata.conversion.build_r_data`,
-    and :func:`rdata.unparser.unparse_file`,
+    This is a convenience function that wraps conversion and unparsing
     as it is the common use case.
 
     Args:
@@ -104,12 +99,13 @@ def write_rda(
         >>> data = {"name": "hello", "values": [1, 2.2, 3.3+4.4j]}
         >>> rdata.write_rda("test.rda", data)
     """
-    r_object = ConverterFromPythonToR(encoding=encoding).convert_to_r_object_for_rda(data)
-    r_data = build_r_data(
-        r_object,
+    converter = ConverterFromPythonToR(
         encoding=encoding,
         format_version=format_version,
     )
+    r_object = converter.convert_to_r_object_for_rda(data)
+    r_data = converter.build_r_data(r_object)
+
     unparse_file(
         path,
         r_data,
