### 2024-09-12 963a9bc
Convert range to array for old format
