### 2024-09-12 937908b
Set object flag explicitly
