### 2024-09-12 efbb09d
Add tests for different dataframe index types
