### 2024-09-12 32a2cc6
Test converting expanded altrep
