### 2024-09-12 1f4e8d8
Add only non-nil attributes to expanded altrep
