### 2024-09-12 237bc22
Enable general rangeindex in dataframe
