### 2024-09-12 1c458ba
Fix ruff and mypy
