### 2024-09-12 92429ca
Add test for dataframe with different dtypes
