### 2024-09-12 5cf678d
Add conversion of boolean pd arrays
