### 2024-09-12 ddabf65
Add missing conversions
