### 2024-09-12 8652271
Add test for dataframe with NAs
