### 2024-09-12 993e2ed
Add dataframe column transformation for more types
