### 2024-09-12 dc1950d
Fix NA values in dataframes
