+++ rdata/parser/__init__.py
@@ -2,6 +2,7 @@
 
 from ._parser import (
     DEFAULT_ALTREP_MAP as DEFAULT_ALTREP_MAP,
+    R_FLOAT_NA as R_FLOAT_NA,
     R_INT_NA as R_INT_NA,
     CharFlags as CharFlags,
     RData as RData,
