+++ rdata/conversion/to_r.py
@@ -10,6 +10,7 @@ import numpy as np
 import pandas as pd
 
 from rdata.parser import (
+    R_FLOAT_NA,
     R_INT_NA,
     CharFlags,
     RData,
@@ -55,23 +56,61 @@ R_MINIMUM_VERSION_WITH_ENCODING: Final[int] = 3
 R_MINIMUM_VERSION_WITH_ALTREP: Final[int] = 3
 
 
-def create_unicode_array(
-        names: Any,  # noqa: ANN401
+def convert_pd_array_to_np_array(
+        pd_array: Any,
 ) -> npt.NDArray[Any]:
     """
-    Create unicode array from sequence/iterator of strings.
+    Convert pandas array object to numpy array.
 
     Args:
-        names: Strings.
+        pd_array: Pandas array.
 
     Returns:
-        Array.
+        Numpy array.
     """
-    name_list = []
-    for name in names:
-        assert isinstance(name, str)
-        name_list.append(name)
-    return np.array(name_list, dtype=np.dtype("U"))
+    if isinstance(pd_array, pd.arrays.StringArray):
+        return pd_array.to_numpy()
+    elif isinstance(pd_array, (
+        pd.arrays.BooleanArray,
+        pd.arrays.IntegerArray,
+        pd.arrays.FloatingArray,
+    )):
+        if isinstance(pd_array, pd.arrays.BooleanArray):
+            dtype = np.bool_
+            fill_value = True
+        elif isinstance(pd_array, pd.arrays.IntegerArray):
+            dtype = np.int32
+            fill_value = R_INT_NA
+        elif isinstance(pd_array, pd.arrays.FloatingArray):
+            dtype = np.float64
+            fill_value = R_FLOAT_NA
+
+        mask = pd_array.isna()
+        if np.any(mask):
+            data = np.empty(pd_array.shape, dtype=dtype)
+            data[~mask] = pd_array[~mask].to_numpy()
+            data[mask] = fill_value
+            if isinstance(pd_array, pd.arrays.FloatingArray):
+                array = data
+            else:
+                array = np.ma.array(
+                    data=data,
+                    mask=mask,
+                    fill_value=fill_value,
+                )
+        else:
+            array = pd_array.to_numpy()
+        assert array.dtype == dtype
+        return array
+
+    elif isinstance(pd_array, (
+        pd.arrays.NumpyExtensionArray,  # type: ignore [attr-defined]
+    )):
+        array = pd_array.to_numpy()
+        return array
+
+    msg = f"pandas array {type(array)} not implemented"
+    raise NotImplementedError(msg)
 
 
 def build_r_object(
@@ -332,17 +371,23 @@ class ConverterFromPythonToR:
             r_value = [self.convert_to_r_object(el) for el in values]
 
             if isinstance(data, dict):
-                names = create_unicode_array(data.keys())
+                names = np.array(list(data.keys()), dtype=np.dtype("U"))
                 attributes = self.build_r_list({"names": names})
 
         elif isinstance(data, np.ndarray):
             if data.dtype.kind in ["O"]:
-                # This is a special case handling only np.array([None])
-                if data.size != 1 or data[0] is not None:
-                    msg = "general object array not implemented"
-                    raise NotImplementedError(msg)
+                assert data.ndim == 1
                 r_type = RObjectType.STR
-                r_value = [build_r_object(RObjectType.CHAR)]
+                r_value = []
+                for el in data:
+                    if el is None or pd.isna(el):
+                        r_el = build_r_object(RObjectType.CHAR)
+                    elif isinstance(el, str):
+                        r_el = self.convert_to_r_object(el.encode(self.encoding))
+                    else:
+                        msg = "general object array not implemented"
[…]
