### 2024-09-12 b622c9c
Add NA floats to ascii parser
