### 2024-09-12 c2728e2
Add NA floats to ascii unparser
