### 2024-09-13 77d726d
Add test for ascii file with NA string
