### 2024-09-13 9da1a42
Define NA checker function close to the definition
