+++ rdata/parser/_parser.py
@@ -36,6 +36,11 @@ R_INT_NA: Final = -2**31
 R_FLOAT_NA: Final = np.frombuffer(b"\x7f\xf0\x00\x00\x00\x00\x07\xa2", dtype=">f8").astype("=f8")[0]  # noqa: E501
 
 
+def is_float_na(value: float) -> bool:
+    """Check if value is NA value."""
+    return np.array(value).tobytes() == np.array(R_FLOAT_NA).tobytes()
+
+
 @runtime_checkable
 class BinaryFileLike(Protocol):
     """Protocol for binary files."""
