### 2024-09-13 d4049ba
Simplify reference lists
