### 2024-09-13 cb3c487
Simplify creation of R lists
