+++ rdata/conversion/to_r.py
@@ -28,19 +28,11 @@ from . import (
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
-    from typing import Any, Final, Literal, Protocol
+    from typing import Any, Final, Literal
 
     import numpy.typing as npt
 
     Encoding = Literal["utf-8", "cp1252"]
-
-
-    class Converter(Protocol):
-        """Protocol for Py-to-R conversion."""
-
-        def __call__(self, data: Any) -> RObject: # noqa: ANN401
-            """Convert Python object to R object."""
-
 
 
 # Default values for RVersions object
@@ -166,6 +158,34 @@ def build_r_object(
      )
 
 
+def build_r_list(
+        data: list[RObject] | list[tuple[RObject, RObject]],
+) -> RObject:
+    """
+    Build R object representing (named) linked list.
+
+    Args:
+        data: Non-empty list of values or (key, value) pairs.
+
+    Returns:
+        R object.
+    """
+    if len(data) == 0:
+        msg = "data must not be empty"
+        raise ValueError(msg)
+
+    head = data[0]
+    tail = data[1:]
+    if isinstance(head, tuple):
+        tag, car = head
+    else:
+        tag = None
+        car = head
+
+    cdr = build_r_object(RObjectType.NILVALUE) if len(tail) == 0 else build_r_list(tail)
+
+    return build_r_object(RObjectType.LIST, value=(car, cdr), tag=tag)
+
 
 class ConverterFromPythonToR:
     """
@@ -229,52 +249,23 @@ class ConverterFromPythonToR:
         return RData(versions, extra, r_object)
 
 
-    def build_r_list(self,
-            data: Mapping[str, Any] | list[Any],
-            *,
-            convert_value: Converter | None = None,
+    def convert_to_r_attributes(self,
+            data: dict[str, Any],
     ) -> RObject:
         """
-        Build R object representing named linked list.
+        Convert dictionary to R attributes list.
 
         Args:
-            data: Non-empty dictionary or list.
-            convert_value: Function used for converting value to R object
-                (for example, convert_to_r_object).
+            data: Non-empty dictionary.
 
         Returns:
             R object.
         """
-        if convert_value is None:
-            convert_value = self.convert_to_r_object
+        converted = []
+        for key, value in data.items():
+            converted.append((self.build_r_sym(key), self.convert_to_r_object(value)))
 
-        if len(data) == 0:
-            msg = "data must not be empty"
-            raise ValueError(msg)
-
-        if isinstance(data, dict):
-            data = data.copy()
-            key = next(iter(data))
-            tag = self.build_r_sym(key)
-            car = data.pop(key)
-        elif isinstance(data, list):
-            car = data[0]
-            data = data[1:]
-            tag = None
-
-        if not isinstance(car, RObject):
-            car = convert_value(car)
-
-        if len(data) == 0:
-            cdr = build_r_object(RObjectType.NILVALUE)
-        else:
-            cdr = self.build_r_list(data, convert_value=convert_value)
-
-        return build_r_object(
-            RObjectType.LIST,
-            value=(car, cdr),
-            tag=tag,
-            )
+        return build_r_list(converted)
 
 
     def build_r_sym(self,
@@ -321,7 +312,7 @@ class ConverterFromPythonToR:
         if not isinstance(data, dict):
             msg = f"for RDA file, data must be a dictionary, not type {type(data)}"
             raise TypeError(msg)
-        return self.build_r_list(data)
+        return self.convert_to_r_attributes(data)
 
 
     def convert_to_r_object(self,  # noqa: C901, PLR0912, PLR0915
@@ -341,7 +332,7 @@ class ConverterFromPythonToR:
[…]
