### 2024-09-13 87194d4
Simplify creation of RData object
