### 2024-09-13 d9c3be5
Add helper functions for conversion
