### 2024-09-16 80a0c9b
Add test for dataframe with NA and NaN floats
