+++ rdata/conversion/_conversion.py
@@ -449,7 +449,13 @@ def _dataframe_column_transform(source: Any) -> Any:  # noqa: ANN401
         if np.issubdtype(source.dtype, np.integer):
             dtype = pd.Int32Dtype()
         elif np.issubdtype(source.dtype, np.floating):
-            dtype = pd.Float64Dtype()
+            # We return the numpy array here, which keeps
+            # R_FLOAT_NA, np.nan, and other NaNs as they were originally in the file.
+            # Users can then decide if they prefer to interpret
+            # only R_FLOAT_NA or all NaNs as "missing".
+            return source
+            # This would create an array with all NaNs as "missing":
+            # dtype = pd.Float64Dtype()  # noqa: ERA001
         elif np.issubdtype(source.dtype, np.complexfloating):
             # There seems to be no pandas type for complex array
             return source
