### 2024-09-16 e88419e
Remove unused R_INT_MIN
