+++ rdata/tests/test_write.py
@@ -139,14 +139,12 @@ def test_convert_to_r(fname: str, expand_altrep: bool) -> None:  # noqa: FBT001
             r_version_serialized=r_data.versions.serialized,
         )
         if fname in [
-            "test_dataframe_dtypes.rds",
-            "test_dataframe_int_rownames.rds",
-            "test_dataframe_range_rownames.rds",
-            "test_dataframe_rownames.rda",
-            "test_dataframe_dtypes_with_na.rds",
-            "test_dataframe_float_with_na_nan.rds",
+            "test_dataframe.rda",
+            "test_dataframe.rds",
+            "test_dataframe_v3.rda",
+            "test_dataframe_v3.rds",
         ]:
-            converter.df_attr_order = ["names", "class", "row.names"]
+            converter.df_attr_order = ["names", "row.names", "class"]
 
         try:
             new_r_data = converter.convert_to_r_data(py_data, file_type=file_type)
