### 2024-09-16 fd813e2
Move NA values and related functions to a new file
