### 2024-09-16 c63dfe7
Add helper functions for handling NA values
