### 2024-09-16 b8b6948
Add comment on setting mask
