### 2024-09-16 828f9da
Include checking int and float values
