### 2024-09-16 0fe903f
Include ascii format in testing too large ints
