### 2024-09-16 7c60566
Speed up range check
