### 2024-09-16 d450f73
Move duplicated code to a function
