### 2024-09-16 4d3e38f
Speed up range check
