### 2024-09-19 957887c
Simplify the definition of the NA value
