### 2024-09-22 5174536
Apply suggestions from code review
Accepted changes from code review:
- Updated references
- Set Zenodo DOIs to "all versions" DOI
- Add references to NumPy and Pandas

Co-authored-by: Tuomas Rossi <34502776+trossi@users.noreply.github.com>
