+++ paper/paper.bib
@@ -1,8 +1,8 @@
-@misc{diaz-vico+ramos-carreno_2022_scikitdatasets,
+@misc{diaz-vico+ramos-carreno_2023_scikitdatasets,
   title = {{{scikit-datasets}}: {{Scikit-learn-compatible}} Datasets},
   author = {{D{\'i}az-Vico}, David and {Ramos-Carre{\~n}o}, Carlos},
-  year = {2022},
-  month = mar,
+  year = {2023},
+  month = aug,
   doi = {10.5281/zenodo.6383047},
   url = {https://github.com/daviddiazvico/scikit-datasets},
   copyright = {MIT}
@@ -13,7 +13,8 @@
   author = {Fajardo, Otto},
   year = {2024},
   month = jul,
-  doi = {10.5281/zenodo.13132498},
+  publisher = {Zenodo},
+  doi = {10.5281/zenodo.7110169},
   url = {https://github.com/ofajardo/pyreadr}
 }
 
@@ -25,11 +26,42 @@
   url = {https://github.com/rpy2/rpy2}
 }
 
-@software{pandasdevelopmentteam_2020_pandasdev,
-  title = {{{pandas-dev/pandas}}: {{pandas}}},
-  author = {Pandas Development Team},
+@article{harris+_2020_numpy,
+  title = {Array programming with {NumPy}},
+  author = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+            van der Walt and Ralf Gommers and Pauli Virtanen and David
+            Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+            Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+            and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+            Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+            R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+            G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+            Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+            Travis E. Oliphant},
   year = {2020},
-  month = feb,
+  month = sep,
+  journal = {Nature},
+  volume = {585},
+  number = {7825},
+  pages = {357--362},
+  doi = {10.1038/s41586-020-2649-2},
+}
+
+@inproceedings{mckinney_2010_pandas,
+  author = {{W}es {M}c{K}inney},
+  title = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
+  booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
+  pages = {56 - 61},
+  year = {2010},
+  editor = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+  doi = {10.25080/Majora-92bf1922-00a},
+}
+
+@software{pandasdevelopmentteam_2024_pandasdev,
+  title = {{{pandas-dev/pandas}}: {{pandas}}},
+  author = {{The Pandas Development Team}},
+  year = {2024},
+  month = apr,
   publisher = {Zenodo},
   doi = {10.5281/zenodo.3509134},
   url = {https://doi.org/10.5281/zenodo.3509134},
