### 2024-10-02 918ea23
Use constant byte-to-string mapping for conversion
