### 2024-10-02 c3fe17c
Apply suggestions from code review
Co-authored-by: Carlos Ramos Carreño <carlosramosca@hotmail.com>
