### 2024-10-02 da9c940
Raise NotImplementedError for untested code
