### 2024-10-02 3acc180
Separate pandas types to constructor functions
