### 2024-10-02 4557d99
Separate string builders to functions
