### 2024-10-02 3e330a3
Remove unused variable
