### 2024-10-02 0d45965
Convert all built-in types via numpy type
