### 2024-10-02 de5a408
Raise error for non-string dictionary keys
