### 2024-10-02 5d8187c
Raise error for non-string rda variable names
