### 2024-10-02 59f4269
Use shorthand function
