### 2024-10-02 540a59a
Add constructor_dict to helper functions
