+++ rdata/conversion/__init__.py
@@ -25,6 +25,8 @@ from ._conversion import (
     ts_constructor as ts_constructor,
 )
 from .to_r import (
+    DEFAULT_CONSTRUCTOR_DICT as DEFAULT_CONSTRUCTOR_DICT,
+    DEFAULT_FORMAT_VERSION as DEFAULT_FORMAT_VERSION,
     ConverterFromPythonToR as ConverterFromPythonToR,
     convert_python_to_r_data as convert_python_to_r_data,
     convert_python_to_r_object as convert_python_to_r_object,
