### 2024-10-02 45820b4
Add tests for altrep wrap_real attributes
