### 2024-10-03 5eae31e
Add sanity check that testing correct object
