### 2024-10-07 577d96f
Apply suggestions from code review
Co-authored-by: Tuomas Rossi <34502776+trossi@users.noreply.github.com>
