### 2024-10-10 118f469
Update paper/paper.md
Co-authored-by: Tuomas Rossi <34502776+trossi@users.noreply.github.com>
