### 2024-10-11 d4bfa29
Update paper/paper.md
Co-authored-by: Tuomas Rossi <34502776+trossi@users.noreply.github.com>
