### 2024-10-24 b8b2aac
Update paper/paper.md
Co-authored-by: Tuomas Rossi <34502776+trossi@users.noreply.github.com>
