### 2024-11-10 #144 @vnmabus
I already submitted to JOSS. The link to the pre-review issue is here: https://github.com/openjournals/joss-reviews/issues/7442
