### 2025-02-05 #144 @coatless
@lwasser as this package was accepted into pyOpenSci and their paper was accepted into JOSS:

https://github.com/openjournals/joss-reviews/issues/7540

Are we able to close out this issue? It's one of the longest active threads. I'm not seeing anything else required on our end.
