### 2025-02-06 #144 @lwasser
@coatless thank you for following up on these older issues 🚀

for these issues all that you need to do it:

1. If accepted by JOSS - the joss accepted label should be added to the issue alongside pyos-accepted - this will move it on our project board
2. update the metadata yaml at the top of the issue with the joss doi (you can see how other reviews have done this)

then close 🚀 thank you for checkin in on it. every once in a while editors forget those last steps because the harder / more focused work on the review end is done :)
