### 2025-02-21 #144 @lwasser
hey everyone! i just updated the package's JOSS doi and label. this issue can be closed. @vnmabus if you want to join our slack, please email me at leah at pyopensci.org or provide an email that you'd like for the invite. You are also welcome to submit a blog post on your package if you are interested. I'll continue to follow notifications here if you want to move forward with that so we can promote your package on social.

Thank you for your patience with the final steps of this review!
